refactor DMAC tests into separate files, update recordings
