+++ sdk/formrecognizer/azure-ai-formrecognizer/tests/recordings/test_dmac_training.test_build_model.yaml
@@ -1,0 +1,92 @@
+interactions:
+- request:
+    body: 'b''{"modelId": "5e2e17ef-661f-46ee-bb63-262da3eb0567", "description": "a
+      v3 model", "azureBlobSource": {"containerUrl": "container_sas_url"}}'''
+    headers:
+      Accept:
+      - application/json
+      Accept-Encoding:
+      - gzip, deflate
+      Connection:
+      - keep-alive
+      Content-Length:
+      - '329'
+      Content-Type:
+      - application/json
+      User-Agent:
+      - azsdk-python-ai-formrecognizer/3.2.0b3 Python/3.8.5 (Windows-10-10.0.19041-SP0)
+    method: POST
+    uri: https://region.api.cognitive.microsoft.com/formrecognizer/documentModels:build?api-version=2021-09-30-preview
+  response:
+    body:
+      string: ''
+    headers:
+      apim-request-id:
+      - 8240ef68-4300-4fe9-bce8-562cfbe1e69c
+      content-length:
+      - '0'
+      date:
+      - Thu, 09 Dec 2021 23:38:34 GMT
+      operation-location:
+      - https://region.api.cognitive.microsoft.com/formrecognizer/operations/31527274888_8240ef68-4300-4fe9-bce8-562cfbe1e69c?api-version=2021-09-30-preview
+      strict-transport-security:
+      - max-age=31536000; includeSubDomains; preload
+      x-content-type-options:
+      - nosniff
+      x-envoy-upstream-service-time:
+      - '3074'
+    status:
+      code: 202
+      message: Accepted
+- request:
+    body: null
+    headers:
+      Accept:
+      - '*/*'
+      Accept-Encoding:
+      - gzip, deflate
+      Connection:
+      - keep-alive
+      User-Agent:
+      - azsdk-python-ai-formrecognizer/3.2.0b3 Python/3.8.5 (Windows-10-10.0.19041-SP0)
+    method: GET
+    uri: https://region.api.cognitive.microsoft.com/formrecognizer/operations/31527274888_8240ef68-4300-4fe9-bce8-562cfbe1e69c?api-version=2021-09-30-preview
+  response:
+    body:
+      string: '{"operationId": "31527274888_8240ef68-4300-4fe9-bce8-562cfbe1e69c",
+        "kind": "documentModelBuild", "status": "succeeded", "createdDateTime": "2021-12-09T23:38:31Z",
+        "lastUpdatedDateTime": "2021-12-09T23:38:39Z", "resourceLocation": "https://region.api.cognitive.microsoft.com//formrecognizer/documentModels/5e2e17ef-661f-46ee-bb63-262da3eb0567?api-version=2021-09-30-preview",
+        "percentCompleted": 100, "result": {"docTypes": {"5e2e17ef-661f-46ee-bb63-262da3eb0567":
+        {"fieldSchema": {"Merchant": {"type": "string"}, "PhoneNumber": {"type": "string"},
+        "Website": {"type": "string"}, "Email": {"type": "string"}, "PurchaseOrderNumber":
+        {"type": "string"}, "DatedAs": {"type": "string"}, "VendorName": {"type":
+        "string"}, "CompanyName": {"type": "string"}, "CompanyAddress": {"type": "string"},
+        "CompanyPhoneNumber": {"type": "string"}, "Subtotal": {"type": "string"},
+        "Tax": {"type": "string"}, "Total": {"type": "string"}, "Signature": {"type":
+        "string"}, "Quantity": {"type": "number"}, "FullSignature": {"type": "signature"}},
+        "technique": "fixedTemplate-2021-07-30", "modelType": "form", "fieldConfidence":
+        {"CompanyAddress": 0.4, "CompanyName": 0.6, "CompanyPhoneNumber": 0.6, "DatedAs":
+        0.6, "Email": 0.4, "FullSignature": 0, "Merchant": 0.6, "PhoneNumber": 0.6,
+        "PurchaseOrderNumber": 0.6, "Quantity": 0.6, "Signature": 0.95, "Subtotal":
+        0.6, "Tax": 0.6, "Total": 0.6, "VendorName": 0.6, "Website": 0.6}}}, "modelId":
+        "5e2e17ef-661f-46ee-bb63-262da3eb0567", "createdDateTime": "2021-12-09T23:38:39Z",
+        "description": "a v3 model"}}'
+    headers:
+      apim-request-id:
+      - 3b41eb22-c9d0-4dc8-bdb7-c18b22ee45bf
+      content-type:
+      - application/json; charset=utf-8
+      date:
+      - Thu, 09 Dec 2021 23:38:39 GMT
+      strict-transport-security:
+      - max-age=31536000; includeSubDomains; preload
+      transfer-encoding:
+      - chunked
+      x-content-type-options:
+      - nosniff
+      x-envoy-upstream-service-time:
+      - '45'
+    status:
+      code: 200
+      message: OK
+version: 1
