+++ sdk/formrecognizer/azure-ai-formrecognizer/tests/recordings/test_dmac_training.test_build_model_auth_bad_key.yaml
@@ -1,6 +1,6 @@
 interactions:
 - request:
-    body: 'b''{"modelId": "d95941c9-e7bf-481f-9124-cb213d27d216", "azureBlobSource":
+    body: 'b''{"modelId": "ad161e6b-b577-4c6d-a86a-2738849b32b4", "azureBlobSource":
       {"containerUrl": "xx"}}'''
     headers:
       Accept:
@@ -14,7 +14,7 @@ interactions:
       Content-Type:
       - application/json
       User-Agent:
-      - azsdk-python-ai-formrecognizer/3.2.0b1 Python/3.9.0 (Windows-10-10.0.19041-SP0)
+      - azsdk-python-ai-formrecognizer/3.2.0b3 Python/3.8.5 (Windows-10-10.0.19041-SP0)
     method: POST
     uri: https://region.api.cognitive.microsoft.com/formrecognizer/documentModels:build?api-version=2021-09-30-preview
   response:
@@ -24,20 +24,14 @@ interactions:
         an active subscription and use a correct regional API endpoint for your resource."}}'
     headers:
       apim-request-id:
-      - 97ecde94-2a81-4fb5-88aa-9f48bdb812a6
+      - 2b37eac2-9be6-4592-95d3-d93be9718d40
       content-length:
-      - '225'
+      - '224'
       content-type:
       - application/json
       date:
-      - Tue, 14 Sep 2021 23:32:52 GMT
-      strict-transport-security:
-      - max-age=31536000; includeSubDomains; preload
-      www-authenticate:
-      - AzureApiManagementKey realm="https://region.api.cognitive.microsoft.com/formrecognizer",name="Ocp-Apim-Subscription-Key",type="header"
-      x-content-type-options:
-      - nosniff
+      - Thu, 09 Dec 2021 23:38:40 GMT
     status:
       code: 401
-      message: Access Denied
+      message: PermissionDenied
 version: 1
