+++ sdk/formrecognizer/azure-ai-formrecognizer/tests/recordings/test_dmac_training.test_build_model_multipage.yaml
@@ -1,6 +1,6 @@
 interactions:
 - request:
-    body: 'b''{"modelId": "b17d6ef9-fd87-472d-b2dc-17e07e81b990", "azureBlobSource":
+    body: 'b''{"modelId": "caf689d9-b7f3-43e6-91da-63135f4453df", "azureBlobSource":
       {"containerUrl": "container_sas_url"}}'''
     headers:
       Accept:
@@ -10,11 +10,11 @@ interactions:
       Connection:
       - keep-alive
       Content-Length:
-      - '299'
+      - '307'
       Content-Type:
       - application/json
       User-Agent:
-      - azsdk-python-ai-formrecognizer/3.2.0b1 Python/3.9.0 (Windows-10-10.0.19041-SP0)
+      - azsdk-python-ai-formrecognizer/3.2.0b3 Python/3.8.5 (Windows-10-10.0.19041-SP0)
     method: POST
     uri: https://region.api.cognitive.microsoft.com/formrecognizer/documentModels:build?api-version=2021-09-30-preview
   response:
@@ -22,19 +22,19 @@ interactions:
       string: ''
     headers:
       apim-request-id:
-      - d2ced60d-5370-434d-931f-80e1693c4b0a
+      - d9679f44-d054-4378-905a-102ffc32f85f
       content-length:
       - '0'
       date:
-      - Wed, 15 Sep 2021 00:37:57 GMT
+      - Thu, 09 Dec 2021 23:38:52 GMT
       operation-location:
-      - https://region.api.cognitive.microsoft.com/formrecognizer/operations/31534701724_d2ced60d-5370-434d-931f-80e1693c4b0a?api-version=2021-09-30-preview
+      - https://region.api.cognitive.microsoft.com/formrecognizer/operations/31527274868_d9679f44-d054-4378-905a-102ffc32f85f?api-version=2021-09-30-preview
       strict-transport-security:
       - max-age=31536000; includeSubDomains; preload
       x-content-type-options:
       - nosniff
       x-envoy-upstream-service-time:
-      - '2348'
+      - '2087'
     status:
       code: 202
       message: Accepted
@@ -48,15 +48,15 @@ interactions:
       Connection:
       - keep-alive
       User-Agent:
-      - azsdk-python-ai-formrecognizer/3.2.0b1 Python/3.9.0 (Windows-10-10.0.19041-SP0)
+      - azsdk-python-ai-formrecognizer/3.2.0b3 Python/3.8.5 (Windows-10-10.0.19041-SP0)
     method: GET
-    uri: https://region.api.cognitive.microsoft.com/formrecognizer/operations/31534701724_d2ced60d-5370-434d-931f-80e1693c4b0a?api-version=2021-09-30-preview
+    uri: https://region.api.cognitive.microsoft.com/formrecognizer/operations/31527274868_d9679f44-d054-4378-905a-102ffc32f85f?api-version=2021-09-30-preview
   response:
     body:
-      string: '{"operationId": "31534701724_d2ced60d-5370-434d-931f-80e1693c4b0a",
-        "kind": "documentModelBuild", "status": "succeeded", "createdDateTime": "2021-09-15T00:37:55Z",
-        "lastUpdatedDateTime": "2021-09-15T00:37:58Z", "resourceLocation": "https://region.api.cognitive.microsoft.com/formrecognizer/documentModels/b17d6ef9-fd87-472d-b2dc-17e07e81b990?api-version=2021-09-30-preview",
-        "percentCompleted": 100, "result": {"docTypes": {"b17d6ef9-fd87-472d-b2dc-17e07e81b990":
+      string: '{"operationId": "31527274868_d9679f44-d054-4378-905a-102ffc32f85f",
+        "kind": "documentModelBuild", "status": "succeeded", "createdDateTime": "2021-12-09T23:38:51Z",
+        "lastUpdatedDateTime": "2021-12-09T23:38:55Z", "resourceLocation": "https://region.api.cognitive.microsoft.com//formrecognizer/documentModels/caf689d9-b7f3-43e6-91da-63135f4453df?api-version=2021-09-30-preview",
+        "percentCompleted": 100, "result": {"docTypes": {"caf689d9-b7f3-43e6-91da-63135f4453df":
         {"fieldSchema": {"Merchant": {"type": "string"}, "MerchantAddress": {"type":
         "string"}, "MerchantPhoneNumber": {"type": "string"}, "CustomerName": {"type":
         "string"}, "CustomerAddress": {"type": "string"}, "CustomerPhoneNumber": {"type":
@@ -65,19 +65,20 @@ interactions:
         "string"}, "Tip": {"type": "string"}, "Total": {"type": "string"}, "Signature":
         {"type": "string"}, "Merchant2": {"type": "string"}, "Customer2": {"type":
         "string"}, "Total2": {"type": "string"}, "Signature2": {"type": "string"}},
-        "fieldConfidence": {"Customer2": 0.95, "CustomerAddress": 0.95, "CustomerName":
-        0.95, "CustomerPhoneNumber": 0.95, "FirstItem": 0.95, "FirstPrice": 0.95,
-        "FirstQuantity": 0.95, "Merchant": 0, "Merchant2": 0, "MerchantAddress": 0.95,
-        "MerchantPhoneNumber": 0.95, "Signature": 0.95, "Signature2": 0.95, "Subtotal":
-        0.95, "Tax": 0.95, "Tip": 0.95, "Total": 0.95, "Total2": 0.95}}}, "modelId":
-        "b17d6ef9-fd87-472d-b2dc-17e07e81b990", "createdDateTime": "2021-09-15T00:37:58Z"}}'
+        "technique": "fixedTemplate-2021-07-30", "modelType": "form", "fieldConfidence":
+        {"Customer2": 0.95, "CustomerAddress": 0.95, "CustomerName": 0.95, "CustomerPhoneNumber":
+        0.95, "FirstItem": 0.95, "FirstPrice": 0.95, "FirstQuantity": 0.95, "Merchant":
+        0, "Merchant2": 0, "MerchantAddress": 0.95, "MerchantPhoneNumber": 0.95, "Signature":
+        0.95, "Signature2": 0.95, "Subtotal": 0.95, "Tax": 0.95, "Tip": 0.95, "Total":
+        0.95, "Total2": 0.95}}}, "modelId": "caf689d9-b7f3-43e6-91da-63135f4453df",
+        "createdDateTime": "2021-12-09T23:38:55Z"}}'
     headers:
       apim-request-id:
-      - 65819283-1f10-4952-a691-d5243fc7f222
+      - 7821759c-7f63-44f2-9f26-3c3380c914db
       content-type:
       - application/json; charset=utf-8
       date:
-      - Wed, 15 Sep 2021 00:38:02 GMT
+      - Thu, 09 Dec 2021 23:38:58 GMT
       strict-transport-security:
       - max-age=31536000; includeSubDomains; preload
       transfer-encoding:
@@ -85,7 +86,7 @@ interactions:
       x-content-type-options:
       - nosniff
       x-envoy-upstream-service-time:
-      - '48'
+      - '38'
     status:
       code: 200
       message: OK
