+++ sdk/formrecognizer/azure-ai-formrecognizer/tests/recordings/test_dmac_training.test_build_model_nested_schema.yaml
@@ -1,0 +1,84 @@
+interactions:
+- request:
+    body: 'b''{"modelId": "57e5a7db-d1d1-40f5-bfae-27970eeda986", "azureBlobSource":
+      {"containerUrl": "container_sas_url"}}'''
+    headers:
+      Accept:
+      - application/json
+      Accept-Encoding:
+      - gzip, deflate
+      Connection:
+      - keep-alive
+      Content-Length:
+      - '318'
+      Content-Type:
+      - application/json
+      User-Agent:
+      - azsdk-python-ai-formrecognizer/3.2.0b3 Python/3.8.5 (Windows-10-10.0.19041-SP0)
+    method: POST
+    uri: https://region.api.cognitive.microsoft.com/formrecognizer/documentModels:build?api-version=2021-09-30-preview
+  response:
+    body:
+      string: ''
+    headers:
+      apim-request-id:
+      - b5bea2fb-9295-4028-8954-2f2dfd8e811f
+      content-length:
+      - '0'
+      date:
+      - Thu, 09 Dec 2021 23:39:09 GMT
+      operation-location:
+      - https://region.api.cognitive.microsoft.com/formrecognizer/operations/31527274852_b5bea2fb-9295-4028-8954-2f2dfd8e811f?api-version=2021-09-30-preview
+      strict-transport-security:
+      - max-age=31536000; includeSubDomains; preload
+      x-content-type-options:
+      - nosniff
+      x-envoy-upstream-service-time:
+      - '2040'
+    status:
+      code: 202
+      message: Accepted
+- request:
+    body: null
+    headers:
+      Accept:
+      - '*/*'
+      Accept-Encoding:
+      - gzip, deflate
+      Connection:
+      - keep-alive
+      User-Agent:
+      - azsdk-python-ai-formrecognizer/3.2.0b3 Python/3.8.5 (Windows-10-10.0.19041-SP0)
+    method: GET
+    uri: https://region.api.cognitive.microsoft.com/formrecognizer/operations/31527274852_b5bea2fb-9295-4028-8954-2f2dfd8e811f?api-version=2021-09-30-preview
+  response:
+    body:
+      string: '{"operationId": "31527274852_b5bea2fb-9295-4028-8954-2f2dfd8e811f",
+        "kind": "documentModelBuild", "status": "succeeded", "createdDateTime": "2021-12-09T23:39:08Z",
+        "lastUpdatedDateTime": "2021-12-09T23:39:11Z", "resourceLocation": "https://region.api.cognitive.microsoft.com//formrecognizer/documentModels/57e5a7db-d1d1-40f5-bfae-27970eeda986?api-version=2021-09-30-preview",
+        "percentCompleted": 100, "result": {"docTypes": {"57e5a7db-d1d1-40f5-bfae-27970eeda986":
+        {"fieldSchema": {"table": {"type": "array", "items": {"type": "object", "properties":
+        {"Item": {"type": "string"}, "Price": {"type": "string"}, "Tax": {"type":
+        "string"}, "Total": {"type": "string"}}}}}, "technique": "fixedTemplate-2021-07-30",
+        "modelType": "form", "fieldConfidence": {"table": 0.95, "table: Item": 0.95,
+        "table: Price": 0.95, "table: Tax": 0.95, "table: Total": 0.95}}}, "modelId":
+        "57e5a7db-d1d1-40f5-bfae-27970eeda986", "createdDateTime": "2021-12-09T23:39:11Z"}}'
+    headers:
+      apim-request-id:
+      - 4da01026-7f55-4863-bed9-0d606162f191
+      content-type:
+      - application/json; charset=utf-8
+      date:
+      - Thu, 09 Dec 2021 23:39:15 GMT
+      strict-transport-security:
+      - max-age=31536000; includeSubDomains; preload
+      transfer-encoding:
+      - chunked
+      x-content-type-options:
+      - nosniff
+      x-envoy-upstream-service-time:
+      - '45'
+    status:
+      code: 200
+      message: OK
+version: 1
