+++ sdk/formrecognizer/azure-ai-formrecognizer/tests/recordings/test_dmac_training.test_build_model_nested_schema_transform.yaml
@@ -1,0 +1,84 @@
+interactions:
+- request:
+    body: 'b''{"modelId": "4852f416-8369-4ac6-b5c5-4c676c2c4f1b", "azureBlobSource":
+      {"containerUrl": "container_sas_url"}}'''
+    headers:
+      Accept:
+      - application/json
+      Accept-Encoding:
+      - gzip, deflate
+      Connection:
+      - keep-alive
+      Content-Length:
+      - '318'
+      Content-Type:
+      - application/json
+      User-Agent:
+      - azsdk-python-ai-formrecognizer/3.2.0b3 Python/3.8.5 (Windows-10-10.0.19041-SP0)
+    method: POST
+    uri: https://region.api.cognitive.microsoft.com/formrecognizer/documentModels:build?api-version=2021-09-30-preview
+  response:
+    body:
+      string: ''
+    headers:
+      apim-request-id:
+      - 83e52eec-d42a-48ad-a8ea-6a99b6449c32
+      content-length:
+      - '0'
+      date:
+      - Thu, 09 Dec 2021 23:39:18 GMT
+      operation-location:
+      - https://region.api.cognitive.microsoft.com/formrecognizer/operations/31527274843_83e52eec-d42a-48ad-a8ea-6a99b6449c32?api-version=2021-09-30-preview
+      strict-transport-security:
+      - max-age=31536000; includeSubDomains; preload
+      x-content-type-options:
+      - nosniff
+      x-envoy-upstream-service-time:
+      - '2128'
+    status:
+      code: 202
+      message: Accepted
+- request:
+    body: null
+    headers:
+      Accept:
+      - '*/*'
+      Accept-Encoding:
+      - gzip, deflate
+      Connection:
+      - keep-alive
+      User-Agent:
+      - azsdk-python-ai-formrecognizer/3.2.0b3 Python/3.8.5 (Windows-10-10.0.19041-SP0)
+    method: GET
+    uri: https://region.api.cognitive.microsoft.com/formrecognizer/operations/31527274843_83e52eec-d42a-48ad-a8ea-6a99b6449c32?api-version=2021-09-30-preview
+  response:
+    body:
+      string: '{"operationId": "31527274843_83e52eec-d42a-48ad-a8ea-6a99b6449c32",
+        "kind": "documentModelBuild", "status": "succeeded", "createdDateTime": "2021-12-09T23:39:16Z",
+        "lastUpdatedDateTime": "2021-12-09T23:39:20Z", "resourceLocation": "https://region.api.cognitive.microsoft.com//formrecognizer/documentModels/4852f416-8369-4ac6-b5c5-4c676c2c4f1b?api-version=2021-09-30-preview",
+        "percentCompleted": 100, "result": {"docTypes": {"4852f416-8369-4ac6-b5c5-4c676c2c4f1b":
+        {"fieldSchema": {"table": {"type": "array", "items": {"type": "object", "properties":
+        {"Item": {"type": "string"}, "Price": {"type": "string"}, "Tax": {"type":
+        "string"}, "Total": {"type": "string"}}}}}, "technique": "fixedTemplate-2021-07-30",
+        "modelType": "form", "fieldConfidence": {"table": 0.95, "table: Item": 0.95,
+        "table: Price": 0.95, "table: Tax": 0.95, "table: Total": 0.95}}}, "modelId":
+        "4852f416-8369-4ac6-b5c5-4c676c2c4f1b", "createdDateTime": "2021-12-09T23:39:20Z"}}'
+    headers:
+      apim-request-id:
+      - db63576c-5fd1-4954-b044-6a2b2968b42c
+      content-type:
+      - application/json; charset=utf-8
+      date:
+      - Thu, 09 Dec 2021 23:39:23 GMT
+      strict-transport-security:
+      - max-age=31536000; includeSubDomains; preload
+      transfer-encoding:
+      - chunked
+      x-content-type-options:
+      - nosniff
+      x-envoy-upstream-service-time:
+      - '39'
+    status:
+      code: 200
+      message: OK
+version: 1
