+++ sdk/formrecognizer/azure-ai-formrecognizer/tests/recordings/test_dmac_training.test_build_model_poller_metadata.yaml
@@ -1,0 +1,91 @@
+interactions:
+- request:
+    body: 'b''{"modelId": "38fd9249-3118-4b87-ad05-3c88108d30f5", "azureBlobSource":
+      {"containerUrl": "container_sas_url"}}'''
+    headers:
+      Accept:
+      - application/json
+      Accept-Encoding:
+      - gzip, deflate
+      Connection:
+      - keep-alive
+      Content-Length:
+      - '300'
+      Content-Type:
+      - application/json
+      User-Agent:
+      - azsdk-python-ai-formrecognizer/3.2.0b3 Python/3.8.5 (Windows-10-10.0.19041-SP0)
+    method: POST
+    uri: https://region.api.cognitive.microsoft.com/formrecognizer/documentModels:build?api-version=2021-09-30-preview
+  response:
+    body:
+      string: ''
+    headers:
+      apim-request-id:
+      - 64e07a4f-d407-491d-8915-7485bccd7c97
+      content-length:
+      - '0'
+      date:
+      - Thu, 09 Dec 2021 23:39:29 GMT
+      operation-location:
+      - https://region.api.cognitive.microsoft.com/formrecognizer/operations/31527274835_64e07a4f-d407-491d-8915-7485bccd7c97?api-version=2021-09-30-preview
+      strict-transport-security:
+      - max-age=31536000; includeSubDomains; preload
+      x-content-type-options:
+      - nosniff
+      x-envoy-upstream-service-time:
+      - '4547'
+    status:
+      code: 202
+      message: Accepted
+- request:
+    body: null
+    headers:
+      Accept:
+      - '*/*'
+      Accept-Encoding:
+      - gzip, deflate
+      Connection:
+      - keep-alive
+      User-Agent:
+      - azsdk-python-ai-formrecognizer/3.2.0b3 Python/3.8.5 (Windows-10-10.0.19041-SP0)
+    method: GET
+    uri: https://region.api.cognitive.microsoft.com/formrecognizer/operations/31527274835_64e07a4f-d407-491d-8915-7485bccd7c97?api-version=2021-09-30-preview
+  response:
+    body:
+      string: '{"operationId": "31527274835_64e07a4f-d407-491d-8915-7485bccd7c97",
+        "kind": "documentModelBuild", "status": "succeeded", "createdDateTime": "2021-12-09T23:39:25Z",
+        "lastUpdatedDateTime": "2021-12-09T23:39:34Z", "resourceLocation": "https://region.api.cognitive.microsoft.com//formrecognizer/documentModels/38fd9249-3118-4b87-ad05-3c88108d30f5?api-version=2021-09-30-preview",
+        "percentCompleted": 100, "result": {"docTypes": {"38fd9249-3118-4b87-ad05-3c88108d30f5":
+        {"fieldSchema": {"Merchant": {"type": "string"}, "PhoneNumber": {"type": "string"},
+        "Website": {"type": "string"}, "Email": {"type": "string"}, "PurchaseOrderNumber":
+        {"type": "string"}, "DatedAs": {"type": "string"}, "VendorName": {"type":
+        "string"}, "CompanyName": {"type": "string"}, "CompanyAddress": {"type": "string"},
+        "CompanyPhoneNumber": {"type": "string"}, "Subtotal": {"type": "string"},
+        "Tax": {"type": "string"}, "Total": {"type": "string"}, "Signature": {"type":
+        "string"}, "Quantity": {"type": "number"}, "FullSignature": {"type": "signature"}},
+        "technique": "fixedTemplate-2021-07-30", "modelType": "form", "fieldConfidence":
+        {"CompanyAddress": 0.4, "CompanyName": 0.6, "CompanyPhoneNumber": 0.6, "DatedAs":
+        0.6, "Email": 0.4, "FullSignature": 0, "Merchant": 0.6, "PhoneNumber": 0.6,
+        "PurchaseOrderNumber": 0.6, "Quantity": 0.6, "Signature": 0.95, "Subtotal":
+        0.6, "Tax": 0.6, "Total": 0.6, "VendorName": 0.6, "Website": 0.6}}}, "modelId":
+        "38fd9249-3118-4b87-ad05-3c88108d30f5", "createdDateTime": "2021-12-09T23:39:34Z"}}'
+    headers:
+      apim-request-id:
+      - 6277f204-415a-4b43-bb2d-c63ca6c9def8
+      content-type:
+      - application/json; charset=utf-8
+      date:
+      - Thu, 09 Dec 2021 23:39:34 GMT
+      strict-transport-security:
+      - max-age=31536000; includeSubDomains; preload
+      transfer-encoding:
+      - chunked
+      x-content-type-options:
+      - nosniff
+      x-envoy-upstream-service-time:
+      - '37'
+    status:
+      code: 200
+      message: OK
+version: 1
