+++ sdk/formrecognizer/azure-ai-formrecognizer/tests/recordings/test_dmac_training.test_build_model_transform.yaml
@@ -1,0 +1,127 @@
+interactions:
+- request:
+    body: 'b''{"modelId": "261185ae-638d-47fc-8a19-65fa223f2d3a", "azureBlobSource":
+      {"containerUrl": "container_sas_url"}}'''
+    headers:
+      Accept:
+      - application/json
+      Accept-Encoding:
+      - gzip, deflate
+      Connection:
+      - keep-alive
+      Content-Length:
+      - '300'
+      Content-Type:
+      - application/json
+      User-Agent:
+      - azsdk-python-ai-formrecognizer/3.2.0b3 Python/3.8.5 (Windows-10-10.0.19041-SP0)
+    method: POST
+    uri: https://region.api.cognitive.microsoft.com/formrecognizer/documentModels:build?api-version=2021-09-30-preview
+  response:
+    body:
+      string: ''
+    headers:
+      apim-request-id:
+      - 613a3a9e-66f4-43a0-96fc-88d57c384490
+      content-length:
+      - '0'
+      date:
+      - Thu, 09 Dec 2021 23:39:57 GMT
+      operation-location:
+      - https://region.api.cognitive.microsoft.com/formrecognizer/operations/31527274806_613a3a9e-66f4-43a0-96fc-88d57c384490?api-version=2021-09-30-preview
+      strict-transport-security:
+      - max-age=31536000; includeSubDomains; preload
+      x-content-type-options:
+      - nosniff
+      x-envoy-upstream-service-time:
+      - '3240'
+    status:
+      code: 202
+      message: Accepted
+- request:
+    body: null
+    headers:
+      Accept:
+      - '*/*'
+      Accept-Encoding:
+      - gzip, deflate
+      Connection:
+      - keep-alive
+      User-Agent:
+      - azsdk-python-ai-formrecognizer/3.2.0b3 Python/3.8.5 (Windows-10-10.0.19041-SP0)
+    method: GET
+    uri: https://region.api.cognitive.microsoft.com/formrecognizer/operations/31527274806_613a3a9e-66f4-43a0-96fc-88d57c384490?api-version=2021-09-30-preview
+  response:
+    body:
+      string: '{"operationId": "31527274806_613a3a9e-66f4-43a0-96fc-88d57c384490",
+        "kind": "documentModelBuild", "status": "running", "createdDateTime": "2021-12-09T23:39:53Z",
+        "lastUpdatedDateTime": "2021-12-09T23:39:57Z", "resourceLocation": "https://region.api.cognitive.microsoft.com//formrecognizer/documentModels/261185ae-638d-47fc-8a19-65fa223f2d3a?api-version=2021-09-30-preview"}'
+    headers:
+      apim-request-id:
+      - 4f783420-9cca-4bb0-baf6-ddf743cebd1c
+      content-type:
+      - application/json; charset=utf-8
+      date:
+      - Thu, 09 Dec 2021 23:40:01 GMT
+      strict-transport-security:
+      - max-age=31536000; includeSubDomains; preload
+      transfer-encoding:
+      - chunked
+      x-content-type-options:
+      - nosniff
+      x-envoy-upstream-service-time:
+      - '37'
+    status:
+      code: 200
+      message: OK
+- request:
+    body: null
+    headers:
+      Accept:
+      - '*/*'
+      Accept-Encoding:
+      - gzip, deflate
+      Connection:
+      - keep-alive
+      User-Agent:
+      - azsdk-python-ai-formrecognizer/3.2.0b3 Python/3.8.5 (Windows-10-10.0.19041-SP0)
+    method: GET
+    uri: https://region.api.cognitive.microsoft.com/formrecognizer/operations/31527274806_613a3a9e-66f4-43a0-96fc-88d57c384490?api-version=2021-09-30-preview
+  response:
+    body:
+      string: '{"operationId": "31527274806_613a3a9e-66f4-43a0-96fc-88d57c384490",
+        "kind": "documentModelBuild", "status": "succeeded", "createdDateTime": "2021-12-09T23:39:53Z",
+        "lastUpdatedDateTime": "2021-12-09T23:40:02Z", "resourceLocation": "https://region.api.cognitive.microsoft.com//formrecognizer/documentModels/261185ae-638d-47fc-8a19-65fa223f2d3a?api-version=2021-09-30-preview",
+        "percentCompleted": 100, "result": {"docTypes": {"261185ae-638d-47fc-8a19-65fa223f2d3a":
+        {"fieldSchema": {"Merchant": {"type": "string"}, "PhoneNumber": {"type": "string"},
+        "Website": {"type": "string"}, "Email": {"type": "string"}, "PurchaseOrderNumber":
+        {"type": "string"}, "DatedAs": {"type": "string"}, "VendorName": {"type":
+        "string"}, "CompanyName": {"type": "string"}, "CompanyAddress": {"type": "string"},
+        "CompanyPhoneNumber": {"type": "string"}, "Subtotal": {"type": "string"},
+        "Tax": {"type": "string"}, "Total": {"type": "string"}, "Signature": {"type":
+        "string"}, "Quantity": {"type": "number"}, "FullSignature": {"type": "signature"}},
+        "technique": "fixedTemplate-2021-07-30", "modelType": "form", "fieldConfidence":
+        {"CompanyAddress": 0.4, "CompanyName": 0.6, "CompanyPhoneNumber": 0.6, "DatedAs":
+        0.6, "Email": 0.4, "FullSignature": 0, "Merchant": 0.6, "PhoneNumber": 0.6,
+        "PurchaseOrderNumber": 0.6, "Quantity": 0.6, "Signature": 0.95, "Subtotal":
+        0.6, "Tax": 0.6, "Total": 0.6, "VendorName": 0.6, "Website": 0.6}}}, "modelId":
+        "261185ae-638d-47fc-8a19-65fa223f2d3a", "createdDateTime": "2021-12-09T23:40:02Z"}}'
+    headers:
+      apim-request-id:
+      - 56dd70ec-40b5-4f20-a28e-f5e93f4031a4
+      content-type:
+      - application/json; charset=utf-8
+      date:
+      - Thu, 09 Dec 2021 23:40:07 GMT
+      strict-transport-security:
+      - max-age=31536000; includeSubDomains; preload
+      transfer-encoding:
+      - chunked
+      x-content-type-options:
+      - nosniff
+      x-envoy-upstream-service-time:
+      - '33'
+    status:
+      code: 200
+      message: OK
+version: 1
