+++ sdk/formrecognizer/azure-ai-formrecognizer/tests/recordings/test_dmac_training_async.test_build_model_auth_bad_key.yaml
@@ -1,6 +1,6 @@
 interactions:
 - request:
-    body: 'b''{"modelId": "7ce5909b-86e5-42f5-bee2-4aa9c83a85b4", "azureBlobSource":
+    body: 'b''{"modelId": "21510c32-3bf3-4ad9-8490-d5e7bf88d9ee", "azureBlobSource":
       {"containerUrl": "xx"}}'''
     headers:
       Accept:
@@ -10,7 +10,7 @@ interactions:
       Content-Type:
       - application/json
       User-Agent:
-      - azsdk-python-ai-formrecognizer/3.2.0b1 Python/3.9.0 (Windows-10-10.0.19041-SP0)
+      - azsdk-python-ai-formrecognizer/3.2.0b3 Python/3.8.5 (Windows-10-10.0.19041-SP0)
     method: POST
     uri: https://region.api.cognitive.microsoft.com/formrecognizer/documentModels:build?api-version=2021-09-30-preview
   response:
@@ -19,15 +19,12 @@ interactions:
         subscription key or wrong API endpoint. Make sure to provide a valid key for
         an active subscription and use a correct regional API endpoint for your resource."}}'
     headers:
-      apim-request-id: 640b9716-e61d-4f38-a9a8-0231f680260f
-      content-length: '225'
+      apim-request-id: e4552c00-5da2-43ac-924a-571d29f500ea
+      content-length: '224'
       content-type: application/json
-      date: Wed, 15 Sep 2021 00:49:46 GMT
-      strict-transport-security: max-age=31536000; includeSubDomains; preload
-      www-authenticate: AzureApiManagementKey realm="https://region.api.cognitive.microsoft.com/formrecognizer",name="Ocp-Apim-Subscription-Key",type="header"
-      x-content-type-options: nosniff
+      date: Thu, 09 Dec 2021 23:35:53 GMT
     status:
       code: 401
-      message: Access Denied
+      message: PermissionDenied
     url: https://region.api.cognitive.microsoft.com/formrecognizer/documentModels:build?api-version=2021-09-30-preview
 version: 1
