+++ sdk/formrecognizer/azure-ai-formrecognizer/tests/recordings/test_dmac_training_async.test_build_model_encoded_url.yaml
@@ -1,31 +1,31 @@
 interactions:
 - request:
-    body: 'b''{"modelId": "45a670ba-205b-46ca-92e9-c0c4528e350f", "azureBlobSource":
-      {"containerUrl": "container_sas_url", "prefix": "subfolder"}}'''
+    body: 'b''{"modelId": "c65203ba-7152-4046-ac5f-83c6b0ed9b39", "azureBlobSource":
+      {"containerUrl": "https://fakeuri.com/blank%20space"}}'''
     headers:
       Accept:
       - application/json
       Content-Length:
-      - '307'
+      - '125'
       Content-Type:
       - application/json
       User-Agent:
-      - azsdk-python-ai-formrecognizer/3.2.0b1 Python/3.9.0 (Windows-10-10.0.19041-SP0)
+      - azsdk-python-ai-formrecognizer/3.2.0b3 Python/3.8.5 (Windows-10-10.0.19041-SP0)
     method: POST
     uri: https://region.api.cognitive.microsoft.com/formrecognizer/documentModels:build?api-version=2021-09-30-preview
   response:
     body:
-      string: '{"error": {"code": "InvalidRequest", "message": "Invalid request.",
-        "innererror": {"code": "TrainingContentMissing", "message": "Training data
-        is missing: Could not find any training data at the given path."}}}'
+      string: '{"error": {"code": "InvalidArgument", "message": "Invalid argument.",
+        "innererror": {"code": "InvalidContentSourceFormat", "message": "Invalid content
+        source: Source Uri must include container path"}}}'
     headers:
-      apim-request-id: b92f457f-2d96-450e-8046-499791fb8975
+      apim-request-id: 5b2a1250-776a-43a8-8776-1a685a3088d0
       content-type: application/json
-      date: Wed, 15 Sep 2021 00:49:46 GMT
+      date: Thu, 09 Dec 2021 23:36:10 GMT
       strict-transport-security: max-age=31536000; includeSubDomains; preload
       transfer-encoding: chunked
       x-content-type-options: nosniff
-      x-envoy-upstream-service-time: '605'
+      x-envoy-upstream-service-time: '48'
     status:
       code: 400
       message: Bad Request
