+++ sdk/formrecognizer/azure-ai-formrecognizer/tests/recordings/test_dmac_training_async.test_build_model_nested_schema.yaml
@@ -1,0 +1,62 @@
+interactions:
+- request:
+    body: 'b''{"modelId": "9ed4677e-9bdf-4bc2-9054-931b135ae9a9", "azureBlobSource":
+      {"containerUrl": "container_sas_url"}}'''
+    headers:
+      Accept:
+      - application/json
+      Content-Length:
+      - '318'
+      Content-Type:
+      - application/json
+      User-Agent:
+      - azsdk-python-ai-formrecognizer/3.2.0b3 Python/3.8.5 (Windows-10-10.0.19041-SP0)
+    method: POST
+    uri: https://region.api.cognitive.microsoft.com/formrecognizer/documentModels:build?api-version=2021-09-30-preview
+  response:
+    body:
+      string: ''
+    headers:
+      apim-request-id: d1627026-f0a1-4108-8e0d-f398c15c8f33
+      content-length: '0'
+      date: Thu, 09 Dec 2021 23:36:29 GMT
+      operation-location: https://region.api.cognitive.microsoft.com/formrecognizer/operations/31527275012_d1627026-f0a1-4108-8e0d-f398c15c8f33?api-version=2021-09-30-preview
+      strict-transport-security: max-age=31536000; includeSubDomains; preload
+      x-content-type-options: nosniff
+      x-envoy-upstream-service-time: '2173'
+    status:
+      code: 202
+      message: Accepted
+    url: https://region.api.cognitive.microsoft.com/formrecognizer/documentModels:build?api-version=2021-09-30-preview
+- request:
+    body: null
+    headers:
+      User-Agent:
+      - azsdk-python-ai-formrecognizer/3.2.0b3 Python/3.8.5 (Windows-10-10.0.19041-SP0)
+    method: GET
+    uri: https://region.api.cognitive.microsoft.com/formrecognizer/operations/31527275012_d1627026-f0a1-4108-8e0d-f398c15c8f33?api-version=2021-09-30-preview
+  response:
+    body:
+      string: '{"operationId": "31527275012_d1627026-f0a1-4108-8e0d-f398c15c8f33",
+        "kind": "documentModelBuild", "status": "succeeded", "createdDateTime": "2021-12-09T23:36:28Z",
+        "lastUpdatedDateTime": "2021-12-09T23:36:31Z", "resourceLocation": "https://region.api.cognitive.microsoft.com//formrecognizer/documentModels/9ed4677e-9bdf-4bc2-9054-931b135ae9a9?api-version=2021-09-30-preview",
+        "percentCompleted": 100, "result": {"docTypes": {"9ed4677e-9bdf-4bc2-9054-931b135ae9a9":
+        {"fieldSchema": {"table": {"type": "array", "items": {"type": "object", "properties":
+        {"Item": {"type": "string"}, "Price": {"type": "string"}, "Tax": {"type":
+        "string"}, "Total": {"type": "string"}}}}}, "technique": "fixedTemplate-2021-07-30",
+        "modelType": "form", "fieldConfidence": {"table": 0.95, "table: Item": 0.95,
+        "table: Price": 0.95, "table: Tax": 0.95, "table: Total": 0.95}}}, "modelId":
+        "9ed4677e-9bdf-4bc2-9054-931b135ae9a9", "createdDateTime": "2021-12-09T23:36:31Z"}}'
+    headers:
+      apim-request-id: 2755c299-58ec-4ded-a35c-a1d262f5e3d4
+      content-type: application/json; charset=utf-8
+      date: Thu, 09 Dec 2021 23:36:35 GMT
+      strict-transport-security: max-age=31536000; includeSubDomains; preload
+      transfer-encoding: chunked
+      x-content-type-options: nosniff
+      x-envoy-upstream-service-time: '45'
+    status:
+      code: 200
+      message: OK
+    url: https://region.api.cognitive.microsoft.com/formrecognizer/operations/31527275012_d1627026-f0a1-4108-8e0d-f398c15c8f33?api-version=2021-09-30-preview
+version: 1
