+++ sdk/formrecognizer/azure-ai-formrecognizer/tests/recordings/test_dmac_training_async.test_build_model_nested_schema_transform.yaml
@@ -1,0 +1,62 @@
+interactions:
+- request:
+    body: 'b''{"modelId": "26c1c5d4-c83c-466a-8b5e-2710d15f3d3d", "azureBlobSource":
+      {"containerUrl": "container_sas_url"}}'''
+    headers:
+      Accept:
+      - application/json
+      Content-Length:
+      - '318'
+      Content-Type:
+      - application/json
+      User-Agent:
+      - azsdk-python-ai-formrecognizer/3.2.0b3 Python/3.8.5 (Windows-10-10.0.19041-SP0)
+    method: POST
+    uri: https://region.api.cognitive.microsoft.com/formrecognizer/documentModels:build?api-version=2021-09-30-preview
+  response:
+    body:
+      string: ''
+    headers:
+      apim-request-id: 03dfc233-2984-416a-b68a-948a8d8a17d7
+      content-length: '0'
+      date: Thu, 09 Dec 2021 23:36:38 GMT
+      operation-location: https://region.api.cognitive.microsoft.com/formrecognizer/operations/31527275003_03dfc233-2984-416a-b68a-948a8d8a17d7?api-version=2021-09-30-preview
+      strict-transport-security: max-age=31536000; includeSubDomains; preload
+      x-content-type-options: nosniff
+      x-envoy-upstream-service-time: '2257'
+    status:
+      code: 202
+      message: Accepted
+    url: https://region.api.cognitive.microsoft.com/formrecognizer/documentModels:build?api-version=2021-09-30-preview
+- request:
+    body: null
+    headers:
+      User-Agent:
+      - azsdk-python-ai-formrecognizer/3.2.0b3 Python/3.8.5 (Windows-10-10.0.19041-SP0)
+    method: GET
+    uri: https://region.api.cognitive.microsoft.com/formrecognizer/operations/31527275003_03dfc233-2984-416a-b68a-948a8d8a17d7?api-version=2021-09-30-preview
+  response:
+    body:
+      string: '{"operationId": "31527275003_03dfc233-2984-416a-b68a-948a8d8a17d7",
+        "kind": "documentModelBuild", "status": "succeeded", "createdDateTime": "2021-12-09T23:36:36Z",
+        "lastUpdatedDateTime": "2021-12-09T23:36:40Z", "resourceLocation": "https://region.api.cognitive.microsoft.com//formrecognizer/documentModels/26c1c5d4-c83c-466a-8b5e-2710d15f3d3d?api-version=2021-09-30-preview",
+        "percentCompleted": 100, "result": {"docTypes": {"26c1c5d4-c83c-466a-8b5e-2710d15f3d3d":
+        {"fieldSchema": {"table": {"type": "array", "items": {"type": "object", "properties":
+        {"Item": {"type": "string"}, "Price": {"type": "string"}, "Tax": {"type":
+        "string"}, "Total": {"type": "string"}}}}}, "technique": "fixedTemplate-2021-07-30",
+        "modelType": "form", "fieldConfidence": {"table": 0.95, "table: Item": 0.95,
+        "table: Price": 0.95, "table: Tax": 0.95, "table: Total": 0.95}}}, "modelId":
+        "26c1c5d4-c83c-466a-8b5e-2710d15f3d3d", "createdDateTime": "2021-12-09T23:36:40Z"}}'
+    headers:
+      apim-request-id: 15ae59dd-a1aa-4f83-885c-fa9941366e6c
+      content-type: application/json; charset=utf-8
+      date: Thu, 09 Dec 2021 23:36:44 GMT
+      strict-transport-security: max-age=31536000; includeSubDomains; preload
+      transfer-encoding: chunked
+      x-content-type-options: nosniff
+      x-envoy-upstream-service-time: '41'
+    status:
+      code: 200
+      message: OK
+    url: https://region.api.cognitive.microsoft.com/formrecognizer/operations/31527275003_03dfc233-2984-416a-b68a-948a8d8a17d7?api-version=2021-09-30-preview
+version: 1
