+++ sdk/formrecognizer/azure-ai-formrecognizer/tests/recordings/test_dmac_training_async.test_build_model_poller_metadata.yaml
@@ -1,0 +1,69 @@
+interactions:
+- request:
+    body: 'b''{"modelId": "392c33ae-4abe-414b-bbd1-7854386e9a9c", "azureBlobSource":
+      {"containerUrl": "container_sas_url"}}'''
+    headers:
+      Accept:
+      - application/json
+      Content-Length:
+      - '300'
+      Content-Type:
+      - application/json
+      User-Agent:
+      - azsdk-python-ai-formrecognizer/3.2.0b3 Python/3.8.5 (Windows-10-10.0.19041-SP0)
+    method: POST
+    uri: https://region.api.cognitive.microsoft.com/formrecognizer/documentModels:build?api-version=2021-09-30-preview
+  response:
+    body:
+      string: ''
+    headers:
+      apim-request-id: a00b7ab7-81ba-4101-bb8c-6f4fb42cd51a
+      content-length: '0'
+      date: Thu, 09 Dec 2021 23:36:47 GMT
+      operation-location: https://region.api.cognitive.microsoft.com/formrecognizer/operations/31527274995_a00b7ab7-81ba-4101-bb8c-6f4fb42cd51a?api-version=2021-09-30-preview
+      strict-transport-security: max-age=31536000; includeSubDomains; preload
+      x-content-type-options: nosniff
+      x-envoy-upstream-service-time: '2875'
+    status:
+      code: 202
+      message: Accepted
+    url: https://region.api.cognitive.microsoft.com/formrecognizer/documentModels:build?api-version=2021-09-30-preview
+- request:
+    body: null
+    headers:
+      User-Agent:
+      - azsdk-python-ai-formrecognizer/3.2.0b3 Python/3.8.5 (Windows-10-10.0.19041-SP0)
+    method: GET
+    uri: https://region.api.cognitive.microsoft.com/formrecognizer/operations/31527274995_a00b7ab7-81ba-4101-bb8c-6f4fb42cd51a?api-version=2021-09-30-preview
+  response:
+    body:
+      string: '{"operationId": "31527274995_a00b7ab7-81ba-4101-bb8c-6f4fb42cd51a",
+        "kind": "documentModelBuild", "status": "succeeded", "createdDateTime": "2021-12-09T23:36:44Z",
+        "lastUpdatedDateTime": "2021-12-09T23:36:52Z", "resourceLocation": "https://region.api.cognitive.microsoft.com//formrecognizer/documentModels/392c33ae-4abe-414b-bbd1-7854386e9a9c?api-version=2021-09-30-preview",
+        "percentCompleted": 100, "result": {"docTypes": {"392c33ae-4abe-414b-bbd1-7854386e9a9c":
+        {"fieldSchema": {"Merchant": {"type": "string"}, "PhoneNumber": {"type": "string"},
+        "Website": {"type": "string"}, "Email": {"type": "string"}, "PurchaseOrderNumber":
+        {"type": "string"}, "DatedAs": {"type": "string"}, "VendorName": {"type":
+        "string"}, "CompanyName": {"type": "string"}, "CompanyAddress": {"type": "string"},
+        "CompanyPhoneNumber": {"type": "string"}, "Subtotal": {"type": "string"},
+        "Tax": {"type": "string"}, "Total": {"type": "string"}, "Signature": {"type":
+        "string"}, "Quantity": {"type": "number"}, "FullSignature": {"type": "signature"}},
+        "technique": "fixedTemplate-2021-07-30", "modelType": "form", "fieldConfidence":
+        {"CompanyAddress": 0.4, "CompanyName": 0.6, "CompanyPhoneNumber": 0.6, "DatedAs":
+        0.6, "Email": 0.4, "FullSignature": 0, "Merchant": 0.6, "PhoneNumber": 0.6,
+        "PurchaseOrderNumber": 0.6, "Quantity": 0.6, "Signature": 0.95, "Subtotal":
+        0.6, "Tax": 0.6, "Total": 0.6, "VendorName": 0.6, "Website": 0.6}}}, "modelId":
+        "392c33ae-4abe-414b-bbd1-7854386e9a9c", "createdDateTime": "2021-12-09T23:36:52Z"}}'
+    headers:
+      apim-request-id: 76b5a12e-fb2f-48b4-a3bd-e6be7c0d5e5a
+      content-type: application/json; charset=utf-8
+      date: Thu, 09 Dec 2021 23:36:52 GMT
+      strict-transport-security: max-age=31536000; includeSubDomains; preload
+      transfer-encoding: chunked
+      x-content-type-options: nosniff
+      x-envoy-upstream-service-time: '41'
+    status:
+      code: 200
+      message: OK
+    url: https://region.api.cognitive.microsoft.com/formrecognizer/operations/31527274995_a00b7ab7-81ba-4101-bb8c-6f4fb42cd51a?api-version=2021-09-30-preview
+version: 1
