+++ sdk/formrecognizer/azure-ai-formrecognizer/tests/recordings/test_dmac_training_async.test_build_model_polling_interval.yaml
@@ -1,0 +1,136 @@
+interactions:
+- request:
+    body: 'b''{"modelId": "caeac8a9-d361-441b-aff3-da6cbc8a669f", "azureBlobSource":
+      {"containerUrl": "container_sas_url"}}'''
+    headers:
+      Accept:
+      - application/json
+      Content-Length:
+      - '300'
+      Content-Type:
+      - application/json
+      User-Agent:
+      - azsdk-python-ai-formrecognizer/3.2.0b3 Python/3.8.5 (Windows-10-10.0.19041-SP0)
+    method: POST
+    uri: https://region.api.cognitive.microsoft.com/formrecognizer/documentModels:build?api-version=2021-09-30-preview
+  response:
+    body:
+      string: ''
+    headers:
+      apim-request-id: 25964b2d-806a-4534-9ea5-416d60c7cfe5
+      content-length: '0'
+      date: Thu, 09 Dec 2021 23:36:55 GMT
+      operation-location: https://region.api.cognitive.microsoft.com/formrecognizer/operations/31527274987_25964b2d-806a-4534-9ea5-416d60c7cfe5?api-version=2021-09-30-preview
+      strict-transport-security: max-age=31536000; includeSubDomains; preload
+      x-content-type-options: nosniff
+      x-envoy-upstream-service-time: '2628'
+    status:
+      code: 202
+      message: Accepted
+    url: https://region.api.cognitive.microsoft.com/formrecognizer/documentModels:build?api-version=2021-09-30-preview
+- request:
+    body: null
+    headers:
+      User-Agent:
+      - azsdk-python-ai-formrecognizer/3.2.0b3 Python/3.8.5 (Windows-10-10.0.19041-SP0)
+    method: GET
+    uri: https://region.api.cognitive.microsoft.com/formrecognizer/operations/31527274987_25964b2d-806a-4534-9ea5-416d60c7cfe5?api-version=2021-09-30-preview
+  response:
+    body:
+      string: '{"operationId": "31527274987_25964b2d-806a-4534-9ea5-416d60c7cfe5",
+        "kind": "documentModelBuild", "status": "succeeded", "createdDateTime": "2021-12-09T23:36:53Z",
+        "lastUpdatedDateTime": "2021-12-09T23:37:00Z", "resourceLocation": "https://region.api.cognitive.microsoft.com//formrecognizer/documentModels/caeac8a9-d361-441b-aff3-da6cbc8a669f?api-version=2021-09-30-preview",
+        "percentCompleted": 100, "result": {"docTypes": {"caeac8a9-d361-441b-aff3-da6cbc8a669f":
+        {"fieldSchema": {"Merchant": {"type": "string"}, "PhoneNumber": {"type": "string"},
+        "Website": {"type": "string"}, "Email": {"type": "string"}, "PurchaseOrderNumber":
+        {"type": "string"}, "DatedAs": {"type": "string"}, "VendorName": {"type":
+        "string"}, "CompanyName": {"type": "string"}, "CompanyAddress": {"type": "string"},
+        "CompanyPhoneNumber": {"type": "string"}, "Subtotal": {"type": "string"},
+        "Tax": {"type": "string"}, "Total": {"type": "string"}, "Signature": {"type":
+        "string"}, "Quantity": {"type": "number"}, "FullSignature": {"type": "signature"}},
+        "technique": "fixedTemplate-2021-07-30", "modelType": "form", "fieldConfidence":
+        {"CompanyAddress": 0.4, "CompanyName": 0.6, "CompanyPhoneNumber": 0.6, "DatedAs":
+        0.6, "Email": 0.4, "FullSignature": 0, "Merchant": 0.6, "PhoneNumber": 0.6,
+        "PurchaseOrderNumber": 0.6, "Quantity": 0.6, "Signature": 0.95, "Subtotal":
+        0.6, "Tax": 0.6, "Total": 0.6, "VendorName": 0.6, "Website": 0.6}}}, "modelId":
+        "caeac8a9-d361-441b-aff3-da6cbc8a669f", "createdDateTime": "2021-12-09T23:37:00Z"}}'
+    headers:
+      apim-request-id: 7b983a8a-4b3e-41d0-9d05-c36d069a7cd2
+      content-type: application/json; charset=utf-8
+      date: Thu, 09 Dec 2021 23:37:01 GMT
+      strict-transport-security: max-age=31536000; includeSubDomains; preload
+      transfer-encoding: chunked
+      x-content-type-options: nosniff
+      x-envoy-upstream-service-time: '41'
+    status:
+      code: 200
+      message: OK
+    url: https://region.api.cognitive.microsoft.com/formrecognizer/operations/31527274987_25964b2d-806a-4534-9ea5-416d60c7cfe5?api-version=2021-09-30-preview
+- request:
+    body: 'b''{"modelId": "5f701844-076c-43a7-adff-bed98e6dba8c", "azureBlobSource":
+      {"containerUrl": "container_sas_url"}}'''
+    headers:
+      Accept:
+      - application/json
+      Content-Length:
+      - '300'
+      Content-Type:
+      - application/json
+      User-Agent:
+      - azsdk-python-ai-formrecognizer/3.2.0b3 Python/3.8.5 (Windows-10-10.0.19041-SP0)
+    method: POST
+    uri: https://region.api.cognitive.microsoft.com/formrecognizer/documentModels:build?api-version=2021-09-30-preview
+  response:
+    body:
+      string: ''
+    headers:
+      apim-request-id: 3e4f74d0-2315-47c0-8f48-162ee59e9ecb
+      content-length: '0'
+      date: Thu, 09 Dec 2021 23:37:04 GMT
+      operation-location: https://region.api.cognitive.microsoft.com/formrecognizer/operations/31527274978_3e4f74d0-2315-47c0-8f48-162ee59e9ecb?api-version=2021-09-30-preview
+      strict-transport-security: max-age=31536000; includeSubDomains; preload
+      x-content-type-options: nosniff
+      x-envoy-upstream-service-time: '2883'
+    status:
+      code: 202
+      message: Accepted
+    url: https://region.api.cognitive.microsoft.com/formrecognizer/documentModels:build?api-version=2021-09-30-preview
+- request:
+    body: null
+    headers:
+      User-Agent:
+      - azsdk-python-ai-formrecognizer/3.2.0b3 Python/3.8.5 (Windows-10-10.0.19041-SP0)
+    method: GET
+    uri: https://region.api.cognitive.microsoft.com/formrecognizer/operations/31527274978_3e4f74d0-2315-47c0-8f48-162ee59e9ecb?api-version=2021-09-30-preview
+  response:
+    body:
+      string: '{"operationId": "31527274978_3e4f74d0-2315-47c0-8f48-162ee59e9ecb",
+        "kind": "documentModelBuild", "status": "succeeded", "createdDateTime": "2021-12-09T23:37:01Z",
+        "lastUpdatedDateTime": "2021-12-09T23:37:09Z", "resourceLocation": "https://region.api.cognitive.microsoft.com//formrecognizer/documentModels/5f701844-076c-43a7-adff-bed98e6dba8c?api-version=2021-09-30-preview",
+        "percentCompleted": 100, "result": {"docTypes": {"5f701844-076c-43a7-adff-bed98e6dba8c":
+        {"fieldSchema": {"Merchant": {"type": "string"}, "PhoneNumber": {"type": "string"},
+        "Website": {"type": "string"}, "Email": {"type": "string"}, "PurchaseOrderNumber":
+        {"type": "string"}, "DatedAs": {"type": "string"}, "VendorName": {"type":
+        "string"}, "CompanyName": {"type": "string"}, "CompanyAddress": {"type": "string"},
+        "CompanyPhoneNumber": {"type": "string"}, "Subtotal": {"type": "string"},
+        "Tax": {"type": "string"}, "Total": {"type": "string"}, "Signature": {"type":
+        "string"}, "Quantity": {"type": "number"}, "FullSignature": {"type": "signature"}},
+        "technique": "fixedTemplate-2021-07-30", "modelType": "form", "fieldConfidence":
+        {"CompanyAddress": 0.4, "CompanyName": 0.6, "CompanyPhoneNumber": 0.6, "DatedAs":
+        0.6, "Email": 0.4, "FullSignature": 0, "Merchant": 0.6, "PhoneNumber": 0.6,
+        "PurchaseOrderNumber": 0.6, "Quantity": 0.6, "Signature": 0.95, "Subtotal":
+        0.6, "Tax": 0.6, "Total": 0.6, "VendorName": 0.6, "Website": 0.6}}}, "modelId":
+        "5f701844-076c-43a7-adff-bed98e6dba8c", "createdDateTime": "2021-12-09T23:37:09Z"}}'
+    headers:
+      apim-request-id: 4bfc1554-9bcc-4307-a535-4a33879ed2f2
+      content-type: application/json; charset=utf-8
+      date: Thu, 09 Dec 2021 23:37:09 GMT
+      strict-transport-security: max-age=31536000; includeSubDomains; preload
+      transfer-encoding: chunked
+      x-content-type-options: nosniff
+      x-envoy-upstream-service-time: '44'
+    status:
+      code: 200
+      message: OK
+    url: https://region.api.cognitive.microsoft.com/formrecognizer/operations/31527274978_3e4f74d0-2315-47c0-8f48-162ee59e9ecb?api-version=2021-09-30-preview
+version: 1
