+++ sdk/formrecognizer/azure-ai-formrecognizer/tests/recordings/test_dmac_training_async.test_build_model_transform.yaml
@@ -1,0 +1,93 @@
+interactions:
+- request:
+    body: 'b''{"modelId": "5c196b92-7d21-4443-afbe-b4af4f485ec5", "azureBlobSource":
+      {"containerUrl": "container_sas_url"}}'''
+    headers:
+      Accept:
+      - application/json
+      Content-Length:
+      - '300'
+      Content-Type:
+      - application/json
+      User-Agent:
+      - azsdk-python-ai-formrecognizer/3.2.0b3 Python/3.8.5 (Windows-10-10.0.19041-SP0)
+    method: POST
+    uri: https://region.api.cognitive.microsoft.com/formrecognizer/documentModels:build?api-version=2021-09-30-preview
+  response:
+    body:
+      string: ''
+    headers:
+      apim-request-id: e4abc029-6a8e-4c11-b9ca-defb7b82730b
+      content-length: '0'
+      date: Thu, 09 Dec 2021 23:37:13 GMT
+      operation-location: https://region.api.cognitive.microsoft.com/formrecognizer/operations/31527274969_e4abc029-6a8e-4c11-b9ca-defb7b82730b?api-version=2021-09-30-preview
+      strict-transport-security: max-age=31536000; includeSubDomains; preload
+      x-content-type-options: nosniff
+      x-envoy-upstream-service-time: '2373'
+    status:
+      code: 202
+      message: Accepted
+    url: https://region.api.cognitive.microsoft.com/formrecognizer/documentModels:build?api-version=2021-09-30-preview
+- request:
+    body: null
+    headers:
+      User-Agent:
+      - azsdk-python-ai-formrecognizer/3.2.0b3 Python/3.8.5 (Windows-10-10.0.19041-SP0)
+    method: GET
+    uri: https://region.api.cognitive.microsoft.com/formrecognizer/operations/31527274969_e4abc029-6a8e-4c11-b9ca-defb7b82730b?api-version=2021-09-30-preview
+  response:
+    body:
+      string: '{"operationId": "31527274969_e4abc029-6a8e-4c11-b9ca-defb7b82730b",
+        "kind": "documentModelBuild", "status": "running", "createdDateTime": "2021-12-09T23:37:11Z",
+        "lastUpdatedDateTime": "2021-12-09T23:37:13Z", "resourceLocation": "https://region.api.cognitive.microsoft.com//formrecognizer/documentModels/5c196b92-7d21-4443-afbe-b4af4f485ec5?api-version=2021-09-30-preview"}'
+    headers:
+      apim-request-id: cfd3bf68-d3fe-4397-ac49-6bc3b22450a7
+      content-type: application/json; charset=utf-8
+      date: Thu, 09 Dec 2021 23:37:18 GMT
+      strict-transport-security: max-age=31536000; includeSubDomains; preload
+      transfer-encoding: chunked
+      x-content-type-options: nosniff
+      x-envoy-upstream-service-time: '36'
+    status:
+      code: 200
+      message: OK
+    url: https://region.api.cognitive.microsoft.com/formrecognizer/operations/31527274969_e4abc029-6a8e-4c11-b9ca-defb7b82730b?api-version=2021-09-30-preview
+- request:
+    body: null
+    headers:
+      User-Agent:
+      - azsdk-python-ai-formrecognizer/3.2.0b3 Python/3.8.5 (Windows-10-10.0.19041-SP0)
+    method: GET
+    uri: https://region.api.cognitive.microsoft.com/formrecognizer/operations/31527274969_e4abc029-6a8e-4c11-b9ca-defb7b82730b?api-version=2021-09-30-preview
+  response:
+    body:
+      string: '{"operationId": "31527274969_e4abc029-6a8e-4c11-b9ca-defb7b82730b",
+        "kind": "documentModelBuild", "status": "succeeded", "createdDateTime": "2021-12-09T23:37:11Z",
+        "lastUpdatedDateTime": "2021-12-09T23:37:19Z", "resourceLocation": "https://region.api.cognitive.microsoft.com//formrecognizer/documentModels/5c196b92-7d21-4443-afbe-b4af4f485ec5?api-version=2021-09-30-preview",
+        "percentCompleted": 100, "result": {"docTypes": {"5c196b92-7d21-4443-afbe-b4af4f485ec5":
+        {"fieldSchema": {"Merchant": {"type": "string"}, "PhoneNumber": {"type": "string"},
+        "Website": {"type": "string"}, "Email": {"type": "string"}, "PurchaseOrderNumber":
+        {"type": "string"}, "DatedAs": {"type": "string"}, "VendorName": {"type":
+        "string"}, "CompanyName": {"type": "string"}, "CompanyAddress": {"type": "string"},
+        "CompanyPhoneNumber": {"type": "string"}, "Subtotal": {"type": "string"},
+        "Tax": {"type": "string"}, "Total": {"type": "string"}, "Signature": {"type":
+        "string"}, "Quantity": {"type": "number"}, "FullSignature": {"type": "signature"}},
+        "technique": "fixedTemplate-2021-07-30", "modelType": "form", "fieldConfidence":
+        {"CompanyAddress": 0.4, "CompanyName": 0.6, "CompanyPhoneNumber": 0.6, "DatedAs":
+        0.6, "Email": 0.4, "FullSignature": 0, "Merchant": 0.6, "PhoneNumber": 0.6,
+        "PurchaseOrderNumber": 0.6, "Quantity": 0.6, "Signature": 0.95, "Subtotal":
+        0.6, "Tax": 0.6, "Total": 0.6, "VendorName": 0.6, "Website": 0.6}}}, "modelId":
+        "5c196b92-7d21-4443-afbe-b4af4f485ec5", "createdDateTime": "2021-12-09T23:37:18Z"}}'
+    headers:
+      apim-request-id: 019fc6c8-dd00-40b4-b164-da70ccfa4376
+      content-type: application/json; charset=utf-8
+      date: Thu, 09 Dec 2021 23:37:23 GMT
+      strict-transport-security: max-age=31536000; includeSubDomains; preload
+      transfer-encoding: chunked
+      x-content-type-options: nosniff
+      x-envoy-upstream-service-time: '40'
+    status:
+      code: 200
+      message: OK
+    url: https://region.api.cognitive.microsoft.com/formrecognizer/operations/31527274969_e4abc029-6a8e-4c11-b9ca-defb7b82730b?api-version=2021-09-30-preview
+version: 1
