--- sdk/formrecognizer/azure-ai-formrecognizer/tests/recordings/test_training.test_build_model.yaml
@@ -1,90 +1,0 @@
-interactions:
-- request:
-    body: 'b''{"modelId": "4a7c7dde-2ab2-48ba-8dc8-47119979e3e2", "description": "a
-      v3 model", "azureBlobSource": {"containerUrl": "container_sas_url"}}'''
-    headers:
-      Accept:
-      - application/json
-      Accept-Encoding:
-      - gzip, deflate
-      Connection:
-      - keep-alive
-      Content-Length:
-      - '313'
-      Content-Type:
-      - application/json
-      User-Agent:
-      - azsdk-python-ai-formrecognizer/3.2.0b1 Python/3.9.0 (Windows-10-10.0.19041-SP0)
-    method: POST
-    uri: https://region.api.cognitive.microsoft.com/formrecognizer/documentModels:build?api-version=2021-09-30-preview
-  response:
-    body:
-      string: ''
-    headers:
-      apim-request-id:
-      - 82a1bd9b-5222-464f-948e-ab88fdd7522f
-      content-length:
-      - '0'
-      date:
-      - Wed, 15 Sep 2021 00:07:53 GMT
-      operation-location:
-      - https://region.api.cognitive.microsoft.com/formrecognizer/operations/31534703529_82a1bd9b-5222-464f-948e-ab88fdd7522f?api-version=2021-09-30-preview
-      strict-transport-security:
-      - max-age=31536000; includeSubDomains; preload
-      x-content-type-options:
-      - nosniff
-      x-envoy-upstream-service-time:
-      - '2848'
-    status:
-      code: 202
-      message: Accepted
-- request:
-    body: null
-    headers:
-      Accept:
-      - '*/*'
-      Accept-Encoding:
-      - gzip, deflate
-      Connection:
-      - keep-alive
-      User-Agent:
-      - azsdk-python-ai-formrecognizer/3.2.0b1 Python/3.9.0 (Windows-10-10.0.19041-SP0)
-    method: GET
-    uri: https://region.api.cognitive.microsoft.com/formrecognizer/operations/31534703529_82a1bd9b-5222-464f-948e-ab88fdd7522f?api-version=2021-09-30-preview
-  response:
-    body:
-      string: '{"operationId": "31534703529_82a1bd9b-5222-464f-948e-ab88fdd7522f",
-        "kind": "documentModelBuild", "status": "succeeded", "createdDateTime": "2021-09-15T00:07:50Z",
-        "lastUpdatedDateTime": "2021-09-15T00:07:55Z", "resourceLocation": "https://region.api.cognitive.microsoft.com/formrecognizer/documentModels/4a7c7dde-2ab2-48ba-8dc8-47119979e3e2?api-version=2021-09-30-preview",
-        "percentCompleted": 100, "result": {"docTypes": {"4a7c7dde-2ab2-48ba-8dc8-47119979e3e2":
-        {"fieldSchema": {"Merchant": {"type": "string"}, "PhoneNumber": {"type": "string"},
-        "Website": {"type": "string"}, "Email": {"type": "string"}, "PurchaseOrderNumber":
-        {"type": "string"}, "DatedAs": {"type": "string"}, "VendorName": {"type":
-        "string"}, "CompanyName": {"type": "string"}, "CompanyAddress": {"type": "string"},
-        "CompanyPhoneNumber": {"type": "string"}, "Subtotal": {"type": "string"},
-        "Tax": {"type": "string"}, "Total": {"type": "string"}, "Signature": {"type":
-        "string"}, "Quantity": {"type": "number"}}, "fieldConfidence": {"CompanyAddress":
-        0.8, "CompanyName": 0.95, "CompanyPhoneNumber": 0.95, "DatedAs": 0.95, "Email":
-        0.8, "Merchant": 0.95, "PhoneNumber": 0.95, "PurchaseOrderNumber": 0.95, "Quantity":
-        0.95, "Signature": 0.6, "Subtotal": 0.95, "Tax": 0.95, "Total": 0.95, "VendorName":
-        0.95, "Website": 0.95}}}, "modelId": "4a7c7dde-2ab2-48ba-8dc8-47119979e3e2",
-        "createdDateTime": "2021-09-15T00:07:55Z", "description": "a v3 model"}}'
-    headers:
-      apim-request-id:
-      - 15fb7920-a052-470c-b359-a80ff224626f
-      content-type:
-      - application/json; charset=utf-8
-      date:
-      - Wed, 15 Sep 2021 00:07:58 GMT
-      strict-transport-security:
-      - max-age=31536000; includeSubDomains; preload
-      transfer-encoding:
-      - chunked
-      x-content-type-options:
-      - nosniff
-      x-envoy-upstream-service-time:
-      - '49'
-    status:
-      code: 200
-      message: OK
-version: 1
