--- sdk/formrecognizer/azure-ai-formrecognizer/tests/recordings/test_training.test_build_model_nested_schema.yaml
@@ -1,84 +1,0 @@
-interactions:
-- request:
-    body: 'b''{"modelId": "4fa30281-9430-4b15-b99f-0c0f0f41ccec", "azureBlobSource":
-      {"containerUrl": "container_sas_url"}}'''
-    headers:
-      Accept:
-      - application/json
-      Accept-Encoding:
-      - gzip, deflate
-      Connection:
-      - keep-alive
-      Content-Length:
-      - '298'
-      Content-Type:
-      - application/json
-      User-Agent:
-      - azsdk-python-ai-formrecognizer/3.2.0b1 Python/3.9.0 (Windows-10-10.0.19041-SP0)
-    method: POST
-    uri: https://region.api.cognitive.microsoft.com/formrecognizer/documentModels:build?api-version=2021-09-30-preview
-  response:
-    body:
-      string: ''
-    headers:
-      apim-request-id:
-      - 61c8c285-8e2a-4042-9b30-2beb4f33c349
-      content-length:
-      - '0'
-      date:
-      - Wed, 15 Sep 2021 00:24:48 GMT
-      operation-location:
-      - https://region.api.cognitive.microsoft.com/formrecognizer/operations/31534702513_61c8c285-8e2a-4042-9b30-2beb4f33c349?api-version=2021-09-30-preview
-      strict-transport-security:
-      - max-age=31536000; includeSubDomains; preload
-      x-content-type-options:
-      - nosniff
-      x-envoy-upstream-service-time:
-      - '2264'
-    status:
-      code: 202
-      message: Accepted
-- request:
-    body: null
-    headers:
-      Accept:
-      - '*/*'
-      Accept-Encoding:
-      - gzip, deflate
-      Connection:
-      - keep-alive
-      User-Agent:
-      - azsdk-python-ai-formrecognizer/3.2.0b1 Python/3.9.0 (Windows-10-10.0.19041-SP0)
-    method: GET
-    uri: https://region.api.cognitive.microsoft.com/formrecognizer/operations/31534702513_61c8c285-8e2a-4042-9b30-2beb4f33c349?api-version=2021-09-30-preview
-  response:
-    body:
-      string: '{"operationId": "31534702513_61c8c285-8e2a-4042-9b30-2beb4f33c349",
-        "kind": "documentModelBuild", "status": "succeeded", "createdDateTime": "2021-09-15T00:24:46Z",
-        "lastUpdatedDateTime": "2021-09-15T00:24:49Z", "resourceLocation": "https://region.api.cognitive.microsoft.com/formrecognizer/documentModels/4fa30281-9430-4b15-b99f-0c0f0f41ccec?api-version=2021-09-30-preview",
-        "percentCompleted": 100, "result": {"docTypes": {"4fa30281-9430-4b15-b99f-0c0f0f41ccec":
-        {"fieldSchema": {"table": {"type": "array", "items": {"type": "object", "properties":
-        {"Item": {"type": "string"}, "Price": {"type": "string"}, "Tax": {"type":
-        "string"}, "Total": {"type": "string"}}}}}, "fieldConfidence": {"table": 0.95,
-        "table: Item": 0.95, "table: Price": 0.95, "table: Tax": 0.95, "table: Total":
-        0.95}}}, "modelId": "4fa30281-9430-4b15-b99f-0c0f0f41ccec", "createdDateTime":
-        "2021-09-15T00:24:49Z"}}'
-    headers:
-      apim-request-id:
-      - 5f33dafb-3509-499d-b790-4c20941e466d
-      content-type:
-      - application/json; charset=utf-8
-      date:
-      - Wed, 15 Sep 2021 00:24:54 GMT
-      strict-transport-security:
-      - max-age=31536000; includeSubDomains; preload
-      transfer-encoding:
-      - chunked
-      x-content-type-options:
-      - nosniff
-      x-envoy-upstream-service-time:
-      - '42'
-    status:
-      code: 200
-      message: OK
-version: 1
