--- sdk/formrecognizer/azure-ai-formrecognizer/tests/recordings/test_training.test_build_model_nested_schema_transform.yaml
@@ -1,84 +1,0 @@
-interactions:
-- request:
-    body: 'b''{"modelId": "a9fc6055-fb47-47ad-afc5-171d5951784b", "azureBlobSource":
-      {"containerUrl": "container_sas_url"}}'''
-    headers:
-      Accept:
-      - application/json
-      Accept-Encoding:
-      - gzip, deflate
-      Connection:
-      - keep-alive
-      Content-Length:
-      - '298'
-      Content-Type:
-      - application/json
-      User-Agent:
-      - azsdk-python-ai-formrecognizer/3.2.0b1 Python/3.9.0 (Windows-10-10.0.19041-SP0)
-    method: POST
-    uri: https://region.api.cognitive.microsoft.com/formrecognizer/documentModels:build?api-version=2021-09-30-preview
-  response:
-    body:
-      string: ''
-    headers:
-      apim-request-id:
-      - 6c66963c-e5fa-4ef5-bea3-1d2b87d4b7b2
-      content-length:
-      - '0'
-      date:
-      - Wed, 15 Sep 2021 00:38:59 GMT
-      operation-location:
-      - https://region.api.cognitive.microsoft.com/formrecognizer/operations/31534701662_6c66963c-e5fa-4ef5-bea3-1d2b87d4b7b2?api-version=2021-09-30-preview
-      strict-transport-security:
-      - max-age=31536000; includeSubDomains; preload
-      x-content-type-options:
-      - nosniff
-      x-envoy-upstream-service-time:
-      - '2237'
-    status:
-      code: 202
-      message: Accepted
-- request:
-    body: null
-    headers:
-      Accept:
-      - '*/*'
-      Accept-Encoding:
-      - gzip, deflate
-      Connection:
-      - keep-alive
-      User-Agent:
-      - azsdk-python-ai-formrecognizer/3.2.0b1 Python/3.9.0 (Windows-10-10.0.19041-SP0)
-    method: GET
-    uri: https://region.api.cognitive.microsoft.com/formrecognizer/operations/31534701662_6c66963c-e5fa-4ef5-bea3-1d2b87d4b7b2?api-version=2021-09-30-preview
-  response:
-    body:
-      string: '{"operationId": "31534701662_6c66963c-e5fa-4ef5-bea3-1d2b87d4b7b2",
-        "kind": "documentModelBuild", "status": "succeeded", "createdDateTime": "2021-09-15T00:38:57Z",
-        "lastUpdatedDateTime": "2021-09-15T00:39:02Z", "resourceLocation": "https://region.api.cognitive.microsoft.com/formrecognizer/documentModels/a9fc6055-fb47-47ad-afc5-171d5951784b?api-version=2021-09-30-preview",
-        "percentCompleted": 100, "result": {"docTypes": {"a9fc6055-fb47-47ad-afc5-171d5951784b":
-        {"fieldSchema": {"table": {"type": "array", "items": {"type": "object", "properties":
-        {"Item": {"type": "string"}, "Price": {"type": "string"}, "Tax": {"type":
-        "string"}, "Total": {"type": "string"}}}}}, "fieldConfidence": {"table": 0.95,
-        "table: Item": 0.95, "table: Price": 0.95, "table: Tax": 0.95, "table: Total":
-        0.95}}}, "modelId": "a9fc6055-fb47-47ad-afc5-171d5951784b", "createdDateTime":
-        "2021-09-15T00:39:02Z"}}'
-    headers:
-      apim-request-id:
-      - d0e0b968-c1aa-4df9-99c5-7e8ae2fa4420
-      content-type:
-      - application/json; charset=utf-8
-      date:
-      - Wed, 15 Sep 2021 00:39:05 GMT
-      strict-transport-security:
-      - max-age=31536000; includeSubDomains; preload
-      transfer-encoding:
-      - chunked
-      x-content-type-options:
-      - nosniff
-      x-envoy-upstream-service-time:
-      - '40'
-    status:
-      code: 200
-      message: OK
-version: 1
