--- sdk/formrecognizer/azure-ai-formrecognizer/tests/recordings/test_training.test_build_model_polling_interval.yaml
@@ -1,178 +1,0 @@
-interactions:
-- request:
-    body: 'b''{"modelId": "9069bcf8-4c01-4368-9c88-4baae83c1e23", "azureBlobSource":
-      {"containerUrl": "container_sas_url"}}'''
-    headers:
-      Accept:
-      - application/json
-      Accept-Encoding:
-      - gzip, deflate
-      Connection:
-      - keep-alive
-      Content-Length:
-      - '284'
-      Content-Type:
-      - application/json
-      User-Agent:
-      - azsdk-python-ai-formrecognizer/3.2.0b1 Python/3.9.0 (Windows-10-10.0.19041-SP0)
-    method: POST
-    uri: https://region.api.cognitive.microsoft.com/formrecognizer/documentModels:build?api-version=2021-09-30-preview
-  response:
-    body:
-      string: ''
-    headers:
-      apim-request-id:
-      - 3f330225-0e49-4243-b7a7-bbb82c5b451d
-      content-length:
-      - '0'
-      date:
-      - Tue, 14 Sep 2021 23:29:47 GMT
-      operation-location:
-      - https://region.api.cognitive.microsoft.com/formrecognizer/operations/31534705814_3f330225-0e49-4243-b7a7-bbb82c5b451d?api-version=2021-09-30-preview
-      strict-transport-security:
-      - max-age=31536000; includeSubDomains; preload
-      x-content-type-options:
-      - nosniff
-      x-envoy-upstream-service-time:
-      - '2585'
-    status:
-      code: 202
-      message: Accepted
-- request:
-    body: null
-    headers:
-      Accept:
-      - '*/*'
-      Accept-Encoding:
-      - gzip, deflate
-      Connection:
-      - keep-alive
-      User-Agent:
-      - azsdk-python-ai-formrecognizer/3.2.0b1 Python/3.9.0 (Windows-10-10.0.19041-SP0)
-    method: GET
-    uri: https://region.api.cognitive.microsoft.com/formrecognizer/operations/31534705814_3f330225-0e49-4243-b7a7-bbb82c5b451d?api-version=2021-09-30-preview
-  response:
-    body:
-      string: '{"operationId": "31534705814_3f330225-0e49-4243-b7a7-bbb82c5b451d",
-        "kind": "documentModelBuild", "status": "succeeded", "createdDateTime": "2021-09-14T23:29:45Z",
-        "lastUpdatedDateTime": "2021-09-14T23:29:51Z", "resourceLocation": "https://region.api.cognitive.microsoft.com/formrecognizer/documentModels/9069bcf8-4c01-4368-9c88-4baae83c1e23?api-version=2021-09-30-preview",
-        "percentCompleted": 100, "result": {"docTypes": {"9069bcf8-4c01-4368-9c88-4baae83c1e23":
-        {"fieldSchema": {"Merchant": {"type": "string"}, "PhoneNumber": {"type": "string"},
-        "Website": {"type": "string"}, "Email": {"type": "string"}, "PurchaseOrderNumber":
-        {"type": "string"}, "DatedAs": {"type": "string"}, "VendorName": {"type":
-        "string"}, "CompanyName": {"type": "string"}, "CompanyAddress": {"type": "string"},
-        "CompanyPhoneNumber": {"type": "string"}, "Subtotal": {"type": "string"},
-        "Tax": {"type": "string"}, "Total": {"type": "string"}, "Signature": {"type":
-        "string"}, "Quantity": {"type": "number"}}, "fieldConfidence": {"CompanyAddress":
-        0.8, "CompanyName": 0.95, "CompanyPhoneNumber": 0.95, "DatedAs": 0.95, "Email":
-        0.8, "Merchant": 0.95, "PhoneNumber": 0.95, "PurchaseOrderNumber": 0.95, "Quantity":
-        0.95, "Signature": 0.6, "Subtotal": 0.95, "Tax": 0.95, "Total": 0.95, "VendorName":
-        0.95, "Website": 0.95}}}, "modelId": "9069bcf8-4c01-4368-9c88-4baae83c1e23",
-        "createdDateTime": "2021-09-14T23:29:50Z"}}'
-    headers:
-      apim-request-id:
-      - aec5931f-80fa-4f0a-8631-48799858523a
-      content-type:
-      - application/json; charset=utf-8
-      date:
-      - Tue, 14 Sep 2021 23:29:54 GMT
-      strict-transport-security:
-      - max-age=31536000; includeSubDomains; preload
-      transfer-encoding:
-      - chunked
-      x-content-type-options:
-      - nosniff
-      x-envoy-upstream-service-time:
-      - '43'
-    status:
-      code: 200
-      message: OK
-- request:
-    body: 'b''{"modelId": "4baef723-dac8-4487-9405-632d06ee2924", "azureBlobSource":
-      {"containerUrl": "container_sas_url"}}'''
-    headers:
-      Accept:
-      - application/json
-      Accept-Encoding:
-      - gzip, deflate
-      Connection:
-      - keep-alive
-      Content-Length:
-      - '284'
-      Content-Type:
-      - application/json
-      User-Agent:
-      - azsdk-python-ai-formrecognizer/3.2.0b1 Python/3.9.0 (Windows-10-10.0.19041-SP0)
-    method: POST
-    uri: https://region.api.cognitive.microsoft.com/formrecognizer/documentModels:build?api-version=2021-09-30-preview
-  response:
-    body:
-      string: ''
-    headers:
-      apim-request-id:
-      - 6c3616f6-6c27-4739-b5e4-efca27309c04
-      content-length:
-      - '0'
-      date:
-      - Tue, 14 Sep 2021 23:29:56 GMT
-      operation-location:
-      - https://region.api.cognitive.microsoft.com/formrecognizer/operations/31534705805_6c3616f6-6c27-4739-b5e4-efca27309c04?api-version=2021-09-30-preview
-      strict-transport-security:
-      - max-age=31536000; includeSubDomains; preload
-      x-content-type-options:
-      - nosniff
-      x-envoy-upstream-service-time:
-      - '2511'
-    status:
-      code: 202
-      message: Accepted
-- request:
-    body: null
-    headers:
-      Accept:
-      - '*/*'
-      Accept-Encoding:
-      - gzip, deflate
-      Connection:
-      - keep-alive
-      User-Agent:
-      - azsdk-python-ai-formrecognizer/3.2.0b1 Python/3.9.0 (Windows-10-10.0.19041-SP0)
-    method: GET
-    uri: https://region.api.cognitive.microsoft.com/formrecognizer/operations/31534705805_6c3616f6-6c27-4739-b5e4-efca27309c04?api-version=2021-09-30-preview
-  response:
-    body:
-      string: '{"operationId": "31534705805_6c3616f6-6c27-4739-b5e4-efca27309c04",
-        "kind": "documentModelBuild", "status": "succeeded", "createdDateTime": "2021-09-14T23:29:54Z",
-        "lastUpdatedDateTime": "2021-09-14T23:29:58Z", "resourceLocation": "https://region.api.cognitive.microsoft.com/formrecognizer/documentModels/4baef723-dac8-4487-9405-632d06ee2924?api-version=2021-09-30-preview",
-        "percentCompleted": 100, "result": {"docTypes": {"4baef723-dac8-4487-9405-632d06ee2924":
-        {"fieldSchema": {"Merchant": {"type": "string"}, "PhoneNumber": {"type": "string"},
-        "Website": {"type": "string"}, "Email": {"type": "string"}, "PurchaseOrderNumber":
-        {"type": "string"}, "DatedAs": {"type": "string"}, "VendorName": {"type":
-        "string"}, "CompanyName": {"type": "string"}, "CompanyAddress": {"type": "string"},
-        "CompanyPhoneNumber": {"type": "string"}, "Subtotal": {"type": "string"},
-        "Tax": {"type": "string"}, "Total": {"type": "string"}, "Signature": {"type":
-        "string"}, "Quantity": {"type": "number"}}, "fieldConfidence": {"CompanyAddress":
-        0.8, "CompanyName": 0.95, "CompanyPhoneNumber": 0.95, "DatedAs": 0.95, "Email":
-        0.8, "Merchant": 0.95, "PhoneNumber": 0.95, "PurchaseOrderNumber": 0.95, "Quantity":
-        0.95, "Signature": 0.6, "Subtotal": 0.95, "Tax": 0.95, "Total": 0.95, "VendorName":
-        0.95, "Website": 0.95}}}, "modelId": "4baef723-dac8-4487-9405-632d06ee2924",
-        "createdDateTime": "2021-09-14T23:29:58Z"}}'
-    headers:
-      apim-request-id:
-      - fbc350c1-1611-4075-9099-721c0a5c5d65
-      content-type:
-      - application/json; charset=utf-8
-      date:
-      - Tue, 14 Sep 2021 23:30:01 GMT
-      strict-transport-security:
-      - max-age=31536000; includeSubDomains; preload
-      transfer-encoding:
-      - chunked
-      x-content-type-options:
-      - nosniff
-      x-envoy-upstream-service-time:
-      - '49'
-    status:
-      code: 200
-      message: OK
-version: 1
