--- sdk/formrecognizer/azure-ai-formrecognizer/tests/recordings/test_training.test_build_model_transform.yaml
@@ -1,90 +1,0 @@
-interactions:
-- request:
-    body: 'b''{"modelId": "cc5dd74e-4949-4b8e-84d6-b972d0abd417", "azureBlobSource":
-      {"containerUrl": "container_sas_url"}}'''
-    headers:
-      Accept:
-      - application/json
-      Accept-Encoding:
-      - gzip, deflate
-      Connection:
-      - keep-alive
-      Content-Length:
-      - '284'
-      Content-Type:
-      - application/json
-      User-Agent:
-      - azsdk-python-ai-formrecognizer/3.2.0b1 Python/3.9.0 (Windows-10-10.0.19041-SP0)
-    method: POST
-    uri: https://region.api.cognitive.microsoft.com/formrecognizer/documentModels:build?api-version=2021-09-30-preview
-  response:
-    body:
-      string: ''
-    headers:
-      apim-request-id:
-      - 180eee2b-2e7d-42d2-b08f-f39a93245448
-      content-length:
-      - '0'
-      date:
-      - Wed, 15 Sep 2021 00:32:34 GMT
-      operation-location:
-      - https://region.api.cognitive.microsoft.com/formrecognizer/operations/31534702048_180eee2b-2e7d-42d2-b08f-f39a93245448?api-version=2021-09-30-preview
-      strict-transport-security:
-      - max-age=31536000; includeSubDomains; preload
-      x-content-type-options:
-      - nosniff
-      x-envoy-upstream-service-time:
-      - '2827'
-    status:
-      code: 202
-      message: Accepted
-- request:
-    body: null
-    headers:
-      Accept:
-      - '*/*'
-      Accept-Encoding:
-      - gzip, deflate
-      Connection:
-      - keep-alive
-      User-Agent:
-      - azsdk-python-ai-formrecognizer/3.2.0b1 Python/3.9.0 (Windows-10-10.0.19041-SP0)
-    method: GET
-    uri: https://region.api.cognitive.microsoft.com/formrecognizer/operations/31534702048_180eee2b-2e7d-42d2-b08f-f39a93245448?api-version=2021-09-30-preview
-  response:
-    body:
-      string: '{"operationId": "31534702048_180eee2b-2e7d-42d2-b08f-f39a93245448",
-        "kind": "documentModelBuild", "status": "succeeded", "createdDateTime": "2021-09-15T00:32:31Z",
-        "lastUpdatedDateTime": "2021-09-15T00:32:36Z", "resourceLocation": "https://region.api.cognitive.microsoft.com/formrecognizer/documentModels/cc5dd74e-4949-4b8e-84d6-b972d0abd417?api-version=2021-09-30-preview",
-        "percentCompleted": 100, "result": {"docTypes": {"cc5dd74e-4949-4b8e-84d6-b972d0abd417":
-        {"fieldSchema": {"Merchant": {"type": "string"}, "PhoneNumber": {"type": "string"},
-        "Website": {"type": "string"}, "Email": {"type": "string"}, "PurchaseOrderNumber":
-        {"type": "string"}, "DatedAs": {"type": "string"}, "VendorName": {"type":
-        "string"}, "CompanyName": {"type": "string"}, "CompanyAddress": {"type": "string"},
-        "CompanyPhoneNumber": {"type": "string"}, "Subtotal": {"type": "string"},
-        "Tax": {"type": "string"}, "Total": {"type": "string"}, "Signature": {"type":
-        "string"}, "Quantity": {"type": "number"}}, "fieldConfidence": {"CompanyAddress":
-        0.8, "CompanyName": 0.95, "CompanyPhoneNumber": 0.95, "DatedAs": 0.95, "Email":
-        0.8, "Merchant": 0.95, "PhoneNumber": 0.95, "PurchaseOrderNumber": 0.95, "Quantity":
-        0.95, "Signature": 0.6, "Subtotal": 0.95, "Tax": 0.95, "Total": 0.95, "VendorName":
-        0.95, "Website": 0.95}}}, "modelId": "cc5dd74e-4949-4b8e-84d6-b972d0abd417",
-        "createdDateTime": "2021-09-15T00:32:36Z"}}'
-    headers:
-      apim-request-id:
-      - 2f1e98ed-349e-4a53-9928-5f0860a99a70
-      content-type:
-      - application/json; charset=utf-8
-      date:
-      - Wed, 15 Sep 2021 00:32:39 GMT
-      strict-transport-security:
-      - max-age=31536000; includeSubDomains; preload
-      transfer-encoding:
-      - chunked
-      x-content-type-options:
-      - nosniff
-      x-envoy-upstream-service-time:
-      - '45'
-    status:
-      code: 200
-      message: OK
-version: 1
