--- sdk/formrecognizer/azure-ai-formrecognizer/tests/recordings/test_training_async.test_build_model.yaml
@@ -1,68 +1,0 @@
-interactions:
-- request:
-    body: 'b''{"modelId": "9aba61ef-367e-479c-b6e4-4c2435bc8dc0", "description": "a
-      v3 model", "azureBlobSource": {"containerUrl": "container_sas_url"}}'''
-    headers:
-      Accept:
-      - application/json
-      Content-Length:
-      - '313'
-      Content-Type:
-      - application/json
-      User-Agent:
-      - azsdk-python-ai-formrecognizer/3.2.0b1 Python/3.9.0 (Windows-10-10.0.19041-SP0)
-    method: POST
-    uri: https://region.api.cognitive.microsoft.com/formrecognizer/documentModels:build?api-version=2021-09-30-preview
-  response:
-    body:
-      string: ''
-    headers:
-      apim-request-id: e6dc73e4-10b0-44ae-9b11-fdbe75d776ba
-      content-length: '0'
-      date: Wed, 15 Sep 2021 00:49:40 GMT
-      operation-location: https://region.api.cognitive.microsoft.com/formrecognizer/operations/31534701022_e6dc73e4-10b0-44ae-9b11-fdbe75d776ba?api-version=2021-09-30-preview
-      strict-transport-security: max-age=31536000; includeSubDomains; preload
-      x-content-type-options: nosniff
-      x-envoy-upstream-service-time: '2629'
-    status:
-      code: 202
-      message: Accepted
-    url: https://region.api.cognitive.microsoft.com/formrecognizer/documentModels:build?api-version=2021-09-30-preview
-- request:
-    body: null
-    headers:
-      User-Agent:
-      - azsdk-python-ai-formrecognizer/3.2.0b1 Python/3.9.0 (Windows-10-10.0.19041-SP0)
-    method: GET
-    uri: https://region.api.cognitive.microsoft.com/formrecognizer/operations/31534701022_e6dc73e4-10b0-44ae-9b11-fdbe75d776ba?api-version=2021-09-30-preview
-  response:
-    body:
-      string: '{"operationId": "31534701022_e6dc73e4-10b0-44ae-9b11-fdbe75d776ba",
-        "kind": "documentModelBuild", "status": "succeeded", "createdDateTime": "2021-09-15T00:49:38Z",
-        "lastUpdatedDateTime": "2021-09-15T00:49:42Z", "resourceLocation": "https://region.api.cognitive.microsoft.com/formrecognizer/documentModels/9aba61ef-367e-479c-b6e4-4c2435bc8dc0?api-version=2021-09-30-preview",
-        "percentCompleted": 100, "result": {"docTypes": {"9aba61ef-367e-479c-b6e4-4c2435bc8dc0":
-        {"fieldSchema": {"Merchant": {"type": "string"}, "PhoneNumber": {"type": "string"},
-        "Website": {"type": "string"}, "Email": {"type": "string"}, "PurchaseOrderNumber":
-        {"type": "string"}, "DatedAs": {"type": "string"}, "VendorName": {"type":
-        "string"}, "CompanyName": {"type": "string"}, "CompanyAddress": {"type": "string"},
-        "CompanyPhoneNumber": {"type": "string"}, "Subtotal": {"type": "string"},
-        "Tax": {"type": "string"}, "Total": {"type": "string"}, "Signature": {"type":
-        "string"}, "Quantity": {"type": "number"}}, "fieldConfidence": {"CompanyAddress":
-        0.8, "CompanyName": 0.95, "CompanyPhoneNumber": 0.95, "DatedAs": 0.95, "Email":
-        0.8, "Merchant": 0.95, "PhoneNumber": 0.95, "PurchaseOrderNumber": 0.95, "Quantity":
-        0.95, "Signature": 0.6, "Subtotal": 0.95, "Tax": 0.95, "Total": 0.95, "VendorName":
-        0.95, "Website": 0.95}}}, "modelId": "9aba61ef-367e-479c-b6e4-4c2435bc8dc0",
-        "createdDateTime": "2021-09-15T00:49:42Z", "description": "a v3 model"}}'
-    headers:
-      apim-request-id: b380074b-dca3-4867-9c23-95578ec4e017
-      content-type: application/json; charset=utf-8
-      date: Wed, 15 Sep 2021 00:49:45 GMT
-      strict-transport-security: max-age=31536000; includeSubDomains; preload
-      transfer-encoding: chunked
-      x-content-type-options: nosniff
-      x-envoy-upstream-service-time: '46'
-    status:
-      code: 200
-      message: OK
-    url: https://region.api.cognitive.microsoft.com/formrecognizer/operations/31534701022_e6dc73e4-10b0-44ae-9b11-fdbe75d776ba?api-version=2021-09-30-preview
-version: 1
