--- sdk/formrecognizer/azure-ai-formrecognizer/tests/recordings/test_training_async.test_build_model_continuation_token.yaml
@@ -1,105 +1,0 @@
-interactions:
-- request:
-    body: 'b''{"modelId": "45754e0e-0aee-4caa-8da2-49ad26dc5ffc", "azureBlobSource":
-      {"containerUrl": "container_sas_url"}}'''
-    headers:
-      Accept:
-      - application/json
-      Content-Length:
-      - '284'
-      Content-Type:
-      - application/json
-      User-Agent:
-      - azsdk-python-ai-formrecognizer/3.2.0b1 Python/3.9.0 (Windows-10-10.0.19041-SP0)
-    method: POST
-    uri: https://region.api.cognitive.microsoft.com/formrecognizer/documentModels:build?api-version=2021-09-30-preview
-  response:
-    body:
-      string: ''
-    headers:
-      apim-request-id: e6e91523-074d-48ff-80b2-c632a1fee8b7
-      content-length: '0'
-      date: Wed, 15 Sep 2021 00:49:49 GMT
-      operation-location: https://region.api.cognitive.microsoft.com/formrecognizer/operations/31534701012_e6e91523-074d-48ff-80b2-c632a1fee8b7?api-version=2021-09-30-preview
-      strict-transport-security: max-age=31536000; includeSubDomains; preload
-      x-content-type-options: nosniff
-      x-envoy-upstream-service-time: '2511'
-    status:
-      code: 202
-      message: Accepted
-    url: https://region.api.cognitive.microsoft.com/formrecognizer/documentModels:build?api-version=2021-09-30-preview
-- request:
-    body: null
-    headers:
-      User-Agent:
-      - azsdk-python-ai-formrecognizer/3.2.0b1 Python/3.9.0 (Windows-10-10.0.19041-SP0)
-    method: GET
-    uri: https://region.api.cognitive.microsoft.com/formrecognizer/operations/31534701012_e6e91523-074d-48ff-80b2-c632a1fee8b7?api-version=2021-09-30-preview
-  response:
-    body:
-      string: '{"operationId": "31534701012_e6e91523-074d-48ff-80b2-c632a1fee8b7",
-        "kind": "documentModelBuild", "status": "succeeded", "createdDateTime": "2021-09-15T00:49:47Z",
-        "lastUpdatedDateTime": "2021-09-15T00:49:51Z", "resourceLocation": "https://region.api.cognitive.microsoft.com/formrecognizer/documentModels/45754e0e-0aee-4caa-8da2-49ad26dc5ffc?api-version=2021-09-30-preview",
-        "percentCompleted": 100, "result": {"docTypes": {"45754e0e-0aee-4caa-8da2-49ad26dc5ffc":
-        {"fieldSchema": {"Merchant": {"type": "string"}, "PhoneNumber": {"type": "string"},
-        "Website": {"type": "string"}, "Email": {"type": "string"}, "PurchaseOrderNumber":
-        {"type": "string"}, "DatedAs": {"type": "string"}, "VendorName": {"type":
-        "string"}, "CompanyName": {"type": "string"}, "CompanyAddress": {"type": "string"},
-        "CompanyPhoneNumber": {"type": "string"}, "Subtotal": {"type": "string"},
-        "Tax": {"type": "string"}, "Total": {"type": "string"}, "Signature": {"type":
-        "string"}, "Quantity": {"type": "number"}}, "fieldConfidence": {"CompanyAddress":
-        0.8, "CompanyName": 0.95, "CompanyPhoneNumber": 0.95, "DatedAs": 0.95, "Email":
-        0.8, "Merchant": 0.95, "PhoneNumber": 0.95, "PurchaseOrderNumber": 0.95, "Quantity":
-        0.95, "Signature": 0.6, "Subtotal": 0.95, "Tax": 0.95, "Total": 0.95, "VendorName":
-        0.95, "Website": 0.95}}}, "modelId": "45754e0e-0aee-4caa-8da2-49ad26dc5ffc",
-        "createdDateTime": "2021-09-15T00:49:51Z"}}'
-    headers:
-      apim-request-id: 26284d82-3489-456e-9cd1-7c030ce50b3c
-      content-type: application/json; charset=utf-8
-      date: Wed, 15 Sep 2021 00:49:55 GMT
-      strict-transport-security: max-age=31536000; includeSubDomains; preload
-      transfer-encoding: chunked
-      x-content-type-options: nosniff
-      x-envoy-upstream-service-time: '40'
-    status:
-      code: 200
-      message: OK
-    url: https://region.api.cognitive.microsoft.com/formrecognizer/operations/31534701012_e6e91523-074d-48ff-80b2-c632a1fee8b7?api-version=2021-09-30-preview
-- request:
-    body: null
-    headers:
-      User-Agent:
-      - azsdk-python-ai-formrecognizer/3.2.0b1 Python/3.9.0 (Windows-10-10.0.19041-SP0)
-    method: GET
-    uri: https://region.api.cognitive.microsoft.com/formrecognizer/operations/31534701012_e6e91523-074d-48ff-80b2-c632a1fee8b7?api-version=2021-09-30-preview
-  response:
-    body:
-      string: '{"operationId": "31534701012_e6e91523-074d-48ff-80b2-c632a1fee8b7",
-        "kind": "documentModelBuild", "status": "succeeded", "createdDateTime": "2021-09-15T00:49:47Z",
-        "lastUpdatedDateTime": "2021-09-15T00:49:51Z", "resourceLocation": "https://region.api.cognitive.microsoft.com/formrecognizer/documentModels/45754e0e-0aee-4caa-8da2-49ad26dc5ffc?api-version=2021-09-30-preview",
-        "percentCompleted": 100, "result": {"docTypes": {"45754e0e-0aee-4caa-8da2-49ad26dc5ffc":
-        {"fieldSchema": {"Merchant": {"type": "string"}, "PhoneNumber": {"type": "string"},
-        "Website": {"type": "string"}, "Email": {"type": "string"}, "PurchaseOrderNumber":
-        {"type": "string"}, "DatedAs": {"type": "string"}, "VendorName": {"type":
-        "string"}, "CompanyName": {"type": "string"}, "CompanyAddress": {"type": "string"},
-        "CompanyPhoneNumber": {"type": "string"}, "Subtotal": {"type": "string"},
-        "Tax": {"type": "string"}, "Total": {"type": "string"}, "Signature": {"type":
-        "string"}, "Quantity": {"type": "number"}}, "fieldConfidence": {"CompanyAddress":
-        0.8, "CompanyName": 0.95, "CompanyPhoneNumber": 0.95, "DatedAs": 0.95, "Email":
-        0.8, "Merchant": 0.95, "PhoneNumber": 0.95, "PurchaseOrderNumber": 0.95, "Quantity":
-        0.95, "Signature": 0.6, "Subtotal": 0.95, "Tax": 0.95, "Total": 0.95, "VendorName":
-        0.95, "Website": 0.95}}}, "modelId": "45754e0e-0aee-4caa-8da2-49ad26dc5ffc",
-        "createdDateTime": "2021-09-15T00:49:51Z"}}'
-    headers:
-      apim-request-id: 1a1c0a7c-da7c-444c-86be-852dbc0aa873
-      content-type: application/json; charset=utf-8
-      date: Wed, 15 Sep 2021 00:50:00 GMT
-      strict-transport-security: max-age=31536000; includeSubDomains; preload
-      transfer-encoding: chunked
-      x-content-type-options: nosniff
-      x-envoy-upstream-service-time: '45'
-    status:
-      code: 200
-      message: OK
-    url: https://region.api.cognitive.microsoft.com/formrecognizer/operations/31534701012_e6e91523-074d-48ff-80b2-c632a1fee8b7?api-version=2021-09-30-preview
-version: 1
