--- sdk/formrecognizer/azure-ai-formrecognizer/tests/recordings/test_training_async.test_build_model_multipage.yaml
@@ -1,70 +1,0 @@
-interactions:
-- request:
-    body: 'b''{"modelId": "9a97e056-22fb-481b-a20d-f642f6e9833a", "azureBlobSource":
-      {"containerUrl": "container_sas_url"}}'''
-    headers:
-      Accept:
-      - application/json
-      Content-Length:
-      - '299'
-      Content-Type:
-      - application/json
-      User-Agent:
-      - azsdk-python-ai-formrecognizer/3.2.0b1 Python/3.9.0 (Windows-10-10.0.19041-SP0)
-    method: POST
-    uri: https://region.api.cognitive.microsoft.com/formrecognizer/documentModels:build?api-version=2021-09-30-preview
-  response:
-    body:
-      string: ''
-    headers:
-      apim-request-id: 935aefbd-6474-485e-aef7-e0ac79c86505
-      content-length: '0'
-      date: Wed, 15 Sep 2021 00:50:03 GMT
-      operation-location: https://region.api.cognitive.microsoft.com/formrecognizer/operations/31534700998_935aefbd-6474-485e-aef7-e0ac79c86505?api-version=2021-09-30-preview
-      strict-transport-security: max-age=31536000; includeSubDomains; preload
-      x-content-type-options: nosniff
-      x-envoy-upstream-service-time: '1839'
-    status:
-      code: 202
-      message: Accepted
-    url: https://region.api.cognitive.microsoft.com/formrecognizer/documentModels:build?api-version=2021-09-30-preview
-- request:
-    body: null
-    headers:
-      User-Agent:
-      - azsdk-python-ai-formrecognizer/3.2.0b1 Python/3.9.0 (Windows-10-10.0.19041-SP0)
-    method: GET
-    uri: https://region.api.cognitive.microsoft.com/formrecognizer/operations/31534700998_935aefbd-6474-485e-aef7-e0ac79c86505?api-version=2021-09-30-preview
-  response:
-    body:
-      string: '{"operationId": "31534700998_935aefbd-6474-485e-aef7-e0ac79c86505",
-        "kind": "documentModelBuild", "status": "succeeded", "createdDateTime": "2021-09-15T00:50:01Z",
-        "lastUpdatedDateTime": "2021-09-15T00:50:04Z", "resourceLocation": "https://region.api.cognitive.microsoft.com/formrecognizer/documentModels/9a97e056-22fb-481b-a20d-f642f6e9833a?api-version=2021-09-30-preview",
-        "percentCompleted": 100, "result": {"docTypes": {"9a97e056-22fb-481b-a20d-f642f6e9833a":
-        {"fieldSchema": {"Merchant": {"type": "string"}, "MerchantAddress": {"type":
-        "string"}, "MerchantPhoneNumber": {"type": "string"}, "CustomerName": {"type":
-        "string"}, "CustomerAddress": {"type": "string"}, "CustomerPhoneNumber": {"type":
-        "string"}, "FirstItem": {"type": "string"}, "FirstQuantity": {"type": "string"},
-        "FirstPrice": {"type": "string"}, "Subtotal": {"type": "string"}, "Tax": {"type":
-        "string"}, "Tip": {"type": "string"}, "Total": {"type": "string"}, "Signature":
-        {"type": "string"}, "Merchant2": {"type": "string"}, "Customer2": {"type":
-        "string"}, "Total2": {"type": "string"}, "Signature2": {"type": "string"}},
-        "fieldConfidence": {"Customer2": 0.95, "CustomerAddress": 0.95, "CustomerName":
-        0.95, "CustomerPhoneNumber": 0.95, "FirstItem": 0.95, "FirstPrice": 0.95,
-        "FirstQuantity": 0.95, "Merchant": 0, "Merchant2": 0, "MerchantAddress": 0.95,
-        "MerchantPhoneNumber": 0.95, "Signature": 0.95, "Signature2": 0.95, "Subtotal":
-        0.95, "Tax": 0.95, "Tip": 0.95, "Total": 0.95, "Total2": 0.95}}}, "modelId":
-        "9a97e056-22fb-481b-a20d-f642f6e9833a", "createdDateTime": "2021-09-15T00:50:04Z"}}'
-    headers:
-      apim-request-id: 7b47040a-35df-4211-8545-4f448763aab1
-      content-type: application/json; charset=utf-8
-      date: Wed, 15 Sep 2021 00:50:08 GMT
-      strict-transport-security: max-age=31536000; includeSubDomains; preload
-      transfer-encoding: chunked
-      x-content-type-options: nosniff
-      x-envoy-upstream-service-time: '39'
-    status:
-      code: 200
-      message: OK
-    url: https://region.api.cognitive.microsoft.com/formrecognizer/operations/31534700998_935aefbd-6474-485e-aef7-e0ac79c86505?api-version=2021-09-30-preview
-version: 1
