--- sdk/formrecognizer/azure-ai-formrecognizer/tests/recordings/test_training_async.test_build_model_multipage_transform.yaml
@@ -1,70 +1,0 @@
-interactions:
-- request:
-    body: 'b''{"modelId": "60eb4c4d-3157-42cd-b8a0-4d943ee4b966", "azureBlobSource":
-      {"containerUrl": "container_sas_url"}}'''
-    headers:
-      Accept:
-      - application/json
-      Content-Length:
-      - '299'
-      Content-Type:
-      - application/json
-      User-Agent:
-      - azsdk-python-ai-formrecognizer/3.2.0b1 Python/3.9.0 (Windows-10-10.0.19041-SP0)
-    method: POST
-    uri: https://region.api.cognitive.microsoft.com/formrecognizer/documentModels:build?api-version=2021-09-30-preview
-  response:
-    body:
-      string: ''
-    headers:
-      apim-request-id: 12c32071-697e-4ac2-87e7-d6fc46897791
-      content-length: '0'
-      date: Wed, 15 Sep 2021 00:50:10 GMT
-      operation-location: https://region.api.cognitive.microsoft.com/formrecognizer/operations/31534700991_12c32071-697e-4ac2-87e7-d6fc46897791?api-version=2021-09-30-preview
-      strict-transport-security: max-age=31536000; includeSubDomains; preload
-      x-content-type-options: nosniff
-      x-envoy-upstream-service-time: '2115'
-    status:
-      code: 202
-      message: Accepted
-    url: https://region.api.cognitive.microsoft.com/formrecognizer/documentModels:build?api-version=2021-09-30-preview
-- request:
-    body: null
-    headers:
-      User-Agent:
-      - azsdk-python-ai-formrecognizer/3.2.0b1 Python/3.9.0 (Windows-10-10.0.19041-SP0)
-    method: GET
-    uri: https://region.api.cognitive.microsoft.com/formrecognizer/operations/31534700991_12c32071-697e-4ac2-87e7-d6fc46897791?api-version=2021-09-30-preview
-  response:
-    body:
-      string: '{"operationId": "31534700991_12c32071-697e-4ac2-87e7-d6fc46897791",
-        "kind": "documentModelBuild", "status": "succeeded", "createdDateTime": "2021-09-15T00:50:09Z",
-        "lastUpdatedDateTime": "2021-09-15T00:50:12Z", "resourceLocation": "https://region.api.cognitive.microsoft.com/formrecognizer/documentModels/60eb4c4d-3157-42cd-b8a0-4d943ee4b966?api-version=2021-09-30-preview",
-        "percentCompleted": 100, "result": {"docTypes": {"60eb4c4d-3157-42cd-b8a0-4d943ee4b966":
-        {"fieldSchema": {"Merchant": {"type": "string"}, "MerchantAddress": {"type":
-        "string"}, "MerchantPhoneNumber": {"type": "string"}, "CustomerName": {"type":
-        "string"}, "CustomerAddress": {"type": "string"}, "CustomerPhoneNumber": {"type":
-        "string"}, "FirstItem": {"type": "string"}, "FirstQuantity": {"type": "string"},
-        "FirstPrice": {"type": "string"}, "Subtotal": {"type": "string"}, "Tax": {"type":
-        "string"}, "Tip": {"type": "string"}, "Total": {"type": "string"}, "Signature":
-        {"type": "string"}, "Merchant2": {"type": "string"}, "Customer2": {"type":
-        "string"}, "Total2": {"type": "string"}, "Signature2": {"type": "string"}},
-        "fieldConfidence": {"Customer2": 0.95, "CustomerAddress": 0.95, "CustomerName":
-        0.95, "CustomerPhoneNumber": 0.95, "FirstItem": 0.95, "FirstPrice": 0.95,
-        "FirstQuantity": 0.95, "Merchant": 0, "Merchant2": 0, "MerchantAddress": 0.95,
-        "MerchantPhoneNumber": 0.95, "Signature": 0.95, "Signature2": 0.95, "Subtotal":
-        0.95, "Tax": 0.95, "Tip": 0.95, "Total": 0.95, "Total2": 0.95}}}, "modelId":
-        "60eb4c4d-3157-42cd-b8a0-4d943ee4b966", "createdDateTime": "2021-09-15T00:50:12Z"}}'
-    headers:
-      apim-request-id: 47f4986b-5a88-4299-8976-00e95408d286
-      content-type: application/json; charset=utf-8
-      date: Wed, 15 Sep 2021 00:50:15 GMT
-      strict-transport-security: max-age=31536000; includeSubDomains; preload
-      transfer-encoding: chunked
-      x-content-type-options: nosniff
-      x-envoy-upstream-service-time: '42'
-    status:
-      code: 200
-      message: OK
-    url: https://region.api.cognitive.microsoft.com/formrecognizer/operations/31534700991_12c32071-697e-4ac2-87e7-d6fc46897791?api-version=2021-09-30-preview
-version: 1
