--- sdk/formrecognizer/azure-ai-formrecognizer/tests/recordings/test_training_async.test_build_model_nested_schema.yaml
@@ -1,62 +1,0 @@
-interactions:
-- request:
-    body: 'b''{"modelId": "159295bc-4410-484b-9796-6fe72c78045e", "azureBlobSource":
-      {"containerUrl": "container_sas_url"}}'''
-    headers:
-      Accept:
-      - application/json
-      Content-Length:
-      - '298'
-      Content-Type:
-      - application/json
-      User-Agent:
-      - azsdk-python-ai-formrecognizer/3.2.0b1 Python/3.9.0 (Windows-10-10.0.19041-SP0)
-    method: POST
-    uri: https://region.api.cognitive.microsoft.com/formrecognizer/documentModels:build?api-version=2021-09-30-preview
-  response:
-    body:
-      string: ''
-    headers:
-      apim-request-id: 192077c6-77e0-404d-9a86-43a1616255d9
-      content-length: '0'
-      date: Wed, 15 Sep 2021 00:50:18 GMT
-      operation-location: https://region.api.cognitive.microsoft.com/formrecognizer/operations/31534700983_192077c6-77e0-404d-9a86-43a1616255d9?api-version=2021-09-30-preview
-      strict-transport-security: max-age=31536000; includeSubDomains; preload
-      x-content-type-options: nosniff
-      x-envoy-upstream-service-time: '1925'
-    status:
-      code: 202
-      message: Accepted
-    url: https://region.api.cognitive.microsoft.com/formrecognizer/documentModels:build?api-version=2021-09-30-preview
-- request:
-    body: null
-    headers:
-      User-Agent:
-      - azsdk-python-ai-formrecognizer/3.2.0b1 Python/3.9.0 (Windows-10-10.0.19041-SP0)
-    method: GET
-    uri: https://region.api.cognitive.microsoft.com/formrecognizer/operations/31534700983_192077c6-77e0-404d-9a86-43a1616255d9?api-version=2021-09-30-preview
-  response:
-    body:
-      string: '{"operationId": "31534700983_192077c6-77e0-404d-9a86-43a1616255d9",
-        "kind": "documentModelBuild", "status": "succeeded", "createdDateTime": "2021-09-15T00:50:16Z",
-        "lastUpdatedDateTime": "2021-09-15T00:50:19Z", "resourceLocation": "https://region.api.cognitive.microsoft.com/formrecognizer/documentModels/159295bc-4410-484b-9796-6fe72c78045e?api-version=2021-09-30-preview",
-        "percentCompleted": 100, "result": {"docTypes": {"159295bc-4410-484b-9796-6fe72c78045e":
-        {"fieldSchema": {"table": {"type": "array", "items": {"type": "object", "properties":
-        {"Item": {"type": "string"}, "Price": {"type": "string"}, "Tax": {"type":
-        "string"}, "Total": {"type": "string"}}}}}, "fieldConfidence": {"table": 0.95,
-        "table: Item": 0.95, "table: Price": 0.95, "table: Tax": 0.95, "table: Total":
-        0.95}}}, "modelId": "159295bc-4410-484b-9796-6fe72c78045e", "createdDateTime":
-        "2021-09-15T00:50:19Z"}}'
-    headers:
-      apim-request-id: 02935291-4c32-45f6-a348-3f3a46c89d5a
-      content-type: application/json; charset=utf-8
-      date: Wed, 15 Sep 2021 00:50:23 GMT
-      strict-transport-security: max-age=31536000; includeSubDomains; preload
-      transfer-encoding: chunked
-      x-content-type-options: nosniff
-      x-envoy-upstream-service-time: '42'
-    status:
-      code: 200
-      message: OK
-    url: https://region.api.cognitive.microsoft.com/formrecognizer/operations/31534700983_192077c6-77e0-404d-9a86-43a1616255d9?api-version=2021-09-30-preview
-version: 1
