--- sdk/formrecognizer/azure-ai-formrecognizer/tests/recordings/test_training_async.test_build_model_nested_schema_transform.yaml
@@ -1,62 +1,0 @@
-interactions:
-- request:
-    body: 'b''{"modelId": "d1e43740-5589-4c98-9e5e-b9301f1cad91", "azureBlobSource":
-      {"containerUrl": "container_sas_url"}}'''
-    headers:
-      Accept:
-      - application/json
-      Content-Length:
-      - '298'
-      Content-Type:
-      - application/json
-      User-Agent:
-      - azsdk-python-ai-formrecognizer/3.2.0b1 Python/3.9.0 (Windows-10-10.0.19041-SP0)
-    method: POST
-    uri: https://region.api.cognitive.microsoft.com/formrecognizer/documentModels:build?api-version=2021-09-30-preview
-  response:
-    body:
-      string: ''
-    headers:
-      apim-request-id: 3ae4f52e-e777-4bfd-b2ca-79c8dda3cbcd
-      content-length: '0'
-      date: Wed, 15 Sep 2021 00:51:26 GMT
-      operation-location: https://region.api.cognitive.microsoft.com/formrecognizer/operations/31534700915_3ae4f52e-e777-4bfd-b2ca-79c8dda3cbcd?api-version=2021-09-30-preview
-      strict-transport-security: max-age=31536000; includeSubDomains; preload
-      x-content-type-options: nosniff
-      x-envoy-upstream-service-time: '2244'
-    status:
-      code: 202
-      message: Accepted
-    url: https://region.api.cognitive.microsoft.com/formrecognizer/documentModels:build?api-version=2021-09-30-preview
-- request:
-    body: null
-    headers:
-      User-Agent:
-      - azsdk-python-ai-formrecognizer/3.2.0b1 Python/3.9.0 (Windows-10-10.0.19041-SP0)
-    method: GET
-    uri: https://region.api.cognitive.microsoft.com/formrecognizer/operations/31534700915_3ae4f52e-e777-4bfd-b2ca-79c8dda3cbcd?api-version=2021-09-30-preview
-  response:
-    body:
-      string: '{"operationId": "31534700915_3ae4f52e-e777-4bfd-b2ca-79c8dda3cbcd",
-        "kind": "documentModelBuild", "status": "succeeded", "createdDateTime": "2021-09-15T00:51:25Z",
-        "lastUpdatedDateTime": "2021-09-15T00:51:29Z", "resourceLocation": "https://region.api.cognitive.microsoft.com/formrecognizer/documentModels/d1e43740-5589-4c98-9e5e-b9301f1cad91?api-version=2021-09-30-preview",
-        "percentCompleted": 100, "result": {"docTypes": {"d1e43740-5589-4c98-9e5e-b9301f1cad91":
-        {"fieldSchema": {"table": {"type": "array", "items": {"type": "object", "properties":
-        {"Item": {"type": "string"}, "Price": {"type": "string"}, "Tax": {"type":
-        "string"}, "Total": {"type": "string"}}}}}, "fieldConfidence": {"table": 0.95,
-        "table: Item": 0.95, "table: Price": 0.95, "table: Tax": 0.95, "table: Total":
-        0.95}}}, "modelId": "d1e43740-5589-4c98-9e5e-b9301f1cad91", "createdDateTime":
-        "2021-09-15T00:51:28Z"}}'
-    headers:
-      apim-request-id: a859fca3-fc2b-4459-9dca-530edea68d35
-      content-type: application/json; charset=utf-8
-      date: Wed, 15 Sep 2021 00:51:32 GMT
-      strict-transport-security: max-age=31536000; includeSubDomains; preload
-      transfer-encoding: chunked
-      x-content-type-options: nosniff
-      x-envoy-upstream-service-time: '47'
-    status:
-      code: 200
-      message: OK
-    url: https://region.api.cognitive.microsoft.com/formrecognizer/operations/31534700915_3ae4f52e-e777-4bfd-b2ca-79c8dda3cbcd?api-version=2021-09-30-preview
-version: 1
