--- sdk/formrecognizer/azure-ai-formrecognizer/tests/recordings/test_training_async.test_build_model_poller_metadata.yaml
@@ -1,68 +1,0 @@
-interactions:
-- request:
-    body: 'b''{"modelId": "65604b9d-4964-4f1f-9792-25f70e478d50", "azureBlobSource":
-      {"containerUrl": "container_sas_url"}}'''
-    headers:
-      Accept:
-      - application/json
-      Content-Length:
-      - '284'
-      Content-Type:
-      - application/json
-      User-Agent:
-      - azsdk-python-ai-formrecognizer/3.2.0b1 Python/3.9.0 (Windows-10-10.0.19041-SP0)
-    method: POST
-    uri: https://region.api.cognitive.microsoft.com/formrecognizer/documentModels:build?api-version=2021-09-30-preview
-  response:
-    body:
-      string: ''
-    headers:
-      apim-request-id: 493ad30c-550c-4127-9da9-a068083260f6
-      content-length: '0'
-      date: Wed, 15 Sep 2021 00:50:26 GMT
-      operation-location: https://region.api.cognitive.microsoft.com/formrecognizer/operations/31534700975_493ad30c-550c-4127-9da9-a068083260f6?api-version=2021-09-30-preview
-      strict-transport-security: max-age=31536000; includeSubDomains; preload
-      x-content-type-options: nosniff
-      x-envoy-upstream-service-time: '2039'
-    status:
-      code: 202
-      message: Accepted
-    url: https://region.api.cognitive.microsoft.com/formrecognizer/documentModels:build?api-version=2021-09-30-preview
-- request:
-    body: null
-    headers:
-      User-Agent:
-      - azsdk-python-ai-formrecognizer/3.2.0b1 Python/3.9.0 (Windows-10-10.0.19041-SP0)
-    method: GET
-    uri: https://region.api.cognitive.microsoft.com/formrecognizer/operations/31534700975_493ad30c-550c-4127-9da9-a068083260f6?api-version=2021-09-30-preview
-  response:
-    body:
-      string: '{"operationId": "31534700975_493ad30c-550c-4127-9da9-a068083260f6",
-        "kind": "documentModelBuild", "status": "succeeded", "createdDateTime": "2021-09-15T00:50:24Z",
-        "lastUpdatedDateTime": "2021-09-15T00:50:28Z", "resourceLocation": "https://region.api.cognitive.microsoft.com/formrecognizer/documentModels/65604b9d-4964-4f1f-9792-25f70e478d50?api-version=2021-09-30-preview",
-        "percentCompleted": 100, "result": {"docTypes": {"65604b9d-4964-4f1f-9792-25f70e478d50":
-        {"fieldSchema": {"Merchant": {"type": "string"}, "PhoneNumber": {"type": "string"},
-        "Website": {"type": "string"}, "Email": {"type": "string"}, "PurchaseOrderNumber":
-        {"type": "string"}, "DatedAs": {"type": "string"}, "VendorName": {"type":
-        "string"}, "CompanyName": {"type": "string"}, "CompanyAddress": {"type": "string"},
-        "CompanyPhoneNumber": {"type": "string"}, "Subtotal": {"type": "string"},
-        "Tax": {"type": "string"}, "Total": {"type": "string"}, "Signature": {"type":
-        "string"}, "Quantity": {"type": "number"}}, "fieldConfidence": {"CompanyAddress":
-        0.8, "CompanyName": 0.95, "CompanyPhoneNumber": 0.95, "DatedAs": 0.95, "Email":
-        0.8, "Merchant": 0.95, "PhoneNumber": 0.95, "PurchaseOrderNumber": 0.95, "Quantity":
-        0.95, "Signature": 0.6, "Subtotal": 0.95, "Tax": 0.95, "Total": 0.95, "VendorName":
-        0.95, "Website": 0.95}}}, "modelId": "65604b9d-4964-4f1f-9792-25f70e478d50",
-        "createdDateTime": "2021-09-15T00:50:28Z"}}'
-    headers:
-      apim-request-id: d13bbb5c-0ab4-4152-9690-263b5a6a0d1f
-      content-type: application/json; charset=utf-8
-      date: Wed, 15 Sep 2021 00:50:31 GMT
-      strict-transport-security: max-age=31536000; includeSubDomains; preload
-      transfer-encoding: chunked
-      x-content-type-options: nosniff
-      x-envoy-upstream-service-time: '38'
-    status:
-      code: 200
-      message: OK
-    url: https://region.api.cognitive.microsoft.com/formrecognizer/operations/31534700975_493ad30c-550c-4127-9da9-a068083260f6?api-version=2021-09-30-preview
-version: 1
