--- sdk/formrecognizer/azure-ai-formrecognizer/tests/recordings/test_training_async.test_build_model_polling_interval.yaml
@@ -1,134 +1,0 @@
-interactions:
-- request:
-    body: 'b''{"modelId": "da29d1e1-f978-43bb-9739-eea9dc338bc8", "azureBlobSource":
-      {"containerUrl": "container_sas_url"}}'''
-    headers:
-      Accept:
-      - application/json
-      Content-Length:
-      - '284'
-      Content-Type:
-      - application/json
-      User-Agent:
-      - azsdk-python-ai-formrecognizer/3.2.0b1 Python/3.9.0 (Windows-10-10.0.19041-SP0)
-    method: POST
-    uri: https://region.api.cognitive.microsoft.com/formrecognizer/documentModels:build?api-version=2021-09-30-preview
-  response:
-    body:
-      string: ''
-    headers:
-      apim-request-id: 3d0cdeea-d6ee-4825-a0c4-fbf0bbaca244
-      content-length: '0'
-      date: Wed, 15 Sep 2021 00:50:34 GMT
-      operation-location: https://region.api.cognitive.microsoft.com/formrecognizer/operations/31534700967_3d0cdeea-d6ee-4825-a0c4-fbf0bbaca244?api-version=2021-09-30-preview
-      strict-transport-security: max-age=31536000; includeSubDomains; preload
-      x-content-type-options: nosniff
-      x-envoy-upstream-service-time: '2576'
-    status:
-      code: 202
-      message: Accepted
-    url: https://region.api.cognitive.microsoft.com/formrecognizer/documentModels:build?api-version=2021-09-30-preview
-- request:
-    body: null
-    headers:
-      User-Agent:
-      - azsdk-python-ai-formrecognizer/3.2.0b1 Python/3.9.0 (Windows-10-10.0.19041-SP0)
-    method: GET
-    uri: https://region.api.cognitive.microsoft.com/formrecognizer/operations/31534700967_3d0cdeea-d6ee-4825-a0c4-fbf0bbaca244?api-version=2021-09-30-preview
-  response:
-    body:
-      string: '{"operationId": "31534700967_3d0cdeea-d6ee-4825-a0c4-fbf0bbaca244",
-        "kind": "documentModelBuild", "status": "succeeded", "createdDateTime": "2021-09-15T00:50:32Z",
-        "lastUpdatedDateTime": "2021-09-15T00:50:36Z", "resourceLocation": "https://region.api.cognitive.microsoft.com/formrecognizer/documentModels/da29d1e1-f978-43bb-9739-eea9dc338bc8?api-version=2021-09-30-preview",
-        "percentCompleted": 100, "result": {"docTypes": {"da29d1e1-f978-43bb-9739-eea9dc338bc8":
-        {"fieldSchema": {"Merchant": {"type": "string"}, "PhoneNumber": {"type": "string"},
-        "Website": {"type": "string"}, "Email": {"type": "string"}, "PurchaseOrderNumber":
-        {"type": "string"}, "DatedAs": {"type": "string"}, "VendorName": {"type":
-        "string"}, "CompanyName": {"type": "string"}, "CompanyAddress": {"type": "string"},
-        "CompanyPhoneNumber": {"type": "string"}, "Subtotal": {"type": "string"},
-        "Tax": {"type": "string"}, "Total": {"type": "string"}, "Signature": {"type":
-        "string"}, "Quantity": {"type": "number"}}, "fieldConfidence": {"CompanyAddress":
-        0.8, "CompanyName": 0.95, "CompanyPhoneNumber": 0.95, "DatedAs": 0.95, "Email":
-        0.8, "Merchant": 0.95, "PhoneNumber": 0.95, "PurchaseOrderNumber": 0.95, "Quantity":
-        0.95, "Signature": 0.6, "Subtotal": 0.95, "Tax": 0.95, "Total": 0.95, "VendorName":
-        0.95, "Website": 0.95}}}, "modelId": "da29d1e1-f978-43bb-9739-eea9dc338bc8",
-        "createdDateTime": "2021-09-15T00:50:36Z"}}'
-    headers:
-      apim-request-id: 16c769be-be28-48f8-946f-5360815bc4d0
-      content-type: application/json; charset=utf-8
-      date: Wed, 15 Sep 2021 00:50:41 GMT
-      strict-transport-security: max-age=31536000; includeSubDomains; preload
-      transfer-encoding: chunked
-      x-content-type-options: nosniff
-      x-envoy-upstream-service-time: '40'
-    status:
-      code: 200
-      message: OK
-    url: https://region.api.cognitive.microsoft.com/formrecognizer/operations/31534700967_3d0cdeea-d6ee-4825-a0c4-fbf0bbaca244?api-version=2021-09-30-preview
-- request:
-    body: 'b''{"modelId": "4b22bd60-7d71-4d7c-90d9-25fc2afa76c8", "azureBlobSource":
-      {"containerUrl": "container_sas_url"}}'''
-    headers:
-      Accept:
-      - application/json
-      Content-Length:
-      - '284'
-      Content-Type:
-      - application/json
-      User-Agent:
-      - azsdk-python-ai-formrecognizer/3.2.0b1 Python/3.9.0 (Windows-10-10.0.19041-SP0)
-    method: POST
-    uri: https://region.api.cognitive.microsoft.com/formrecognizer/documentModels:build?api-version=2021-09-30-preview
-  response:
-    body:
-      string: ''
-    headers:
-      apim-request-id: 0673652f-5bc8-44a1-8e62-a297242d55d8
-      content-length: '0'
-      date: Wed, 15 Sep 2021 00:50:44 GMT
-      operation-location: https://region.api.cognitive.microsoft.com/formrecognizer/operations/31534700958_0673652f-5bc8-44a1-8e62-a297242d55d8?api-version=2021-09-30-preview
-      strict-transport-security: max-age=31536000; includeSubDomains; preload
-      x-content-type-options: nosniff
-      x-envoy-upstream-service-time: '2953'
-    status:
-      code: 202
-      message: Accepted
-    url: https://region.api.cognitive.microsoft.com/formrecognizer/documentModels:build?api-version=2021-09-30-preview
-- request:
-    body: null
-    headers:
-      User-Agent:
-      - azsdk-python-ai-formrecognizer/3.2.0b1 Python/3.9.0 (Windows-10-10.0.19041-SP0)
-    method: GET
-    uri: https://region.api.cognitive.microsoft.com/formrecognizer/operations/31534700958_0673652f-5bc8-44a1-8e62-a297242d55d8?api-version=2021-09-30-preview
-  response:
-    body:
-      string: '{"operationId": "31534700958_0673652f-5bc8-44a1-8e62-a297242d55d8",
-        "kind": "documentModelBuild", "status": "succeeded", "createdDateTime": "2021-09-15T00:50:41Z",
-        "lastUpdatedDateTime": "2021-09-15T00:50:46Z", "resourceLocation": "https://region.api.cognitive.microsoft.com/formrecognizer/documentModels/4b22bd60-7d71-4d7c-90d9-25fc2afa76c8?api-version=2021-09-30-preview",
-        "percentCompleted": 100, "result": {"docTypes": {"4b22bd60-7d71-4d7c-90d9-25fc2afa76c8":
-        {"fieldSchema": {"Merchant": {"type": "string"}, "PhoneNumber": {"type": "string"},
-        "Website": {"type": "string"}, "Email": {"type": "string"}, "PurchaseOrderNumber":
-        {"type": "string"}, "DatedAs": {"type": "string"}, "VendorName": {"type":
-        "string"}, "CompanyName": {"type": "string"}, "CompanyAddress": {"type": "string"},
-        "CompanyPhoneNumber": {"type": "string"}, "Subtotal": {"type": "string"},
-        "Tax": {"type": "string"}, "Total": {"type": "string"}, "Signature": {"type":
-        "string"}, "Quantity": {"type": "number"}}, "fieldConfidence": {"CompanyAddress":
-        0.8, "CompanyName": 0.95, "CompanyPhoneNumber": 0.95, "DatedAs": 0.95, "Email":
-        0.8, "Merchant": 0.95, "PhoneNumber": 0.95, "PurchaseOrderNumber": 0.95, "Quantity":
-        0.95, "Signature": 0.6, "Subtotal": 0.95, "Tax": 0.95, "Total": 0.95, "VendorName":
-        0.95, "Website": 0.95}}}, "modelId": "4b22bd60-7d71-4d7c-90d9-25fc2afa76c8",
-        "createdDateTime": "2021-09-15T00:50:46Z"}}'
-    headers:
-      apim-request-id: 09521994-797d-491f-9888-fb0fed0f6f17
-      content-type: application/json; charset=utf-8
-      date: Wed, 15 Sep 2021 00:50:49 GMT
-      strict-transport-security: max-age=31536000; includeSubDomains; preload
-      transfer-encoding: chunked
-      x-content-type-options: nosniff
-      x-envoy-upstream-service-time: '45'
-    status:
-      code: 200
-      message: OK
-    url: https://region.api.cognitive.microsoft.com/formrecognizer/operations/31534700958_0673652f-5bc8-44a1-8e62-a297242d55d8?api-version=2021-09-30-preview
-version: 1
